@@ -9,26 +9,23 @@ plugins {
 
 android {
 
-    compileSdkVersion(29)
-    buildToolsVersion("30.0.0-rc1")
+    compileSdkVersion(30)
+    buildToolsVersion("30.0.2")
     defaultConfig {
         applicationId = "it.cammino.risuscito"
         minSdkVersion(16)
-        targetSdkVersion(29)
+        targetSdkVersion(30)
         multiDexEnabled = true
-        versionCode = 4503
-        versionName = "4.5.0"
+        versionCode = 4611
+        versionName = "4.6.0"
         vectorDrawables.useSupportLibrary = true
         javaCompileOptions {
             annotationProcessorOptions {
-                arguments = mapOf(
+                mapOf(
                         "room.schemaLocation" to "$projectDir/schemas",
                         "room.incremental" to "true",
                         "room.expandProjection" to "true")
             }
-        }
-        viewBinding {
-            isEnabled = true
         }
     }
 
