@@ -7,6 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.blogspot.atifsoftwares.animatoolib.Animatoo
+import com.google.android.material.elevation.ElevationOverlayProvider
 import com.vansuita.materialabout.builder.AboutBuilder
 import it.cammino.risuscito.databinding.AboutLayoutBinding
 
@@ -46,21 +47,22 @@ class AboutFragment : Fragment() {
         }
 
         binding.about.addView(
-                AboutBuilder.with(mMainActivity)
-                        .setAppIcon(R.drawable.ic_launcher_144dp)
-                        .setAppName(R.string.app_name)
-                        .setPhoto(R.drawable.ic_brand_icon)
-                        .setCover(R.mipmap.profile_cover)
-                        .setLinksColumnsCount(1)
-                        .addEmailLink("marbat87@outlook.it", getString(R.string.app_name), null)
-                        .addFiveStarsAction(BuildConfig.APPLICATION_ID)
-                        .setVersionNameAsAppSubTitle()
-                        .addShareAction(R.string.app_name)
-                        .addUpdateAction(BuildConfig.APPLICATION_ID)
-                        .setActionsColumnsCount(2)
-                        .addChangeLogAction(mChangeLogClickListener)
-                        .addPrivacyPolicyAction("http://marbat87.altervista.org/privacy_policy.html")
-                        .setShowAsCard(false)
-                        .build())
+                AboutBuilder.with(mMainActivity).apply {
+                    setAppIcon(R.drawable.ic_launcher_144dp)
+                    setAppName(R.string.app_name)
+                    setPhoto(R.drawable.ic_brand_icon)
+                    setCover(R.mipmap.profile_cover)
+                    linksColumnsCount = 1
+                    addEmailLink("marbat87@outlook.it", getString(R.string.app_name), null)
+                    addFiveStarsAction(BuildConfig.APPLICATION_ID)
+                    setVersionNameAsAppSubTitle()
+                    addShareAction(R.string.app_name)
+                    addUpdateAction(BuildConfig.APPLICATION_ID)
+                    actionsColumnsCount = 2
+                    addChangeLogAction(mChangeLogClickListener)
+                    addPrivacyPolicyAction("https://marbat87.altervista.org/privacy_policy.html")
+                    isShowAsCard = false
+                    backgroundColor = ElevationOverlayProvider(requireContext()).compositeOverlayWithThemeSurfaceColorIfNeeded(resources.getDimension(R.dimen.mtrl_card_elevation))
+                }.build())
     }
 }
