@@ -1,14 +1,12 @@
 package it.cammino.risuscito
 
 import android.annotation.SuppressLint
-import android.app.Activity
 import android.graphics.Color
 import android.graphics.Typeface
-import android.os.AsyncTask
+import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
 import android.os.Handler
-import android.os.Message
-import android.text.Editable
+import android.os.Looper
 import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
@@ -23,20 +21,21 @@ import androidx.core.os.postDelayed
 import androidx.core.view.ViewCompat
 import androidx.core.view.isVisible
 import androidx.core.widget.doOnTextChanged
-import androidx.lifecycle.observe
+import androidx.lifecycle.lifecycleScope
 import androidx.preference.PreferenceManager
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.afollestad.materialdialogs.MaterialDialog
-import com.afollestad.materialdialogs.input.getInputField
 import com.blogspot.atifsoftwares.animatoolib.Animatoo
 import com.getkeepsafe.taptargetview.TapTarget
 import com.getkeepsafe.taptargetview.TapTargetSequence
+import com.google.android.material.color.MaterialColors
+import com.google.android.material.elevation.ElevationOverlayProvider
 import com.google.android.material.snackbar.Snackbar
-import com.mikepenz.fastadapter.IAdapter
 import com.mikepenz.fastadapter.adapters.FastItemAdapter
+import com.mikepenz.fastadapter.binding.BindingViewHolder
+import com.mikepenz.fastadapter.binding.listeners.addLongClickListener
 import com.mikepenz.fastadapter.drag.ItemTouchCallback
 import com.mikepenz.fastadapter.swipe.SimpleSwipeCallback
 import com.mikepenz.fastadapter.swipe_drag.SimpleSwipeDragCallback
@@ -50,24 +49,31 @@ import com.mikepenz.iconics.utils.sizeDp
 import it.cammino.risuscito.database.RisuscitoDatabase
 import it.cammino.risuscito.database.entities.ListaPers
 import it.cammino.risuscito.databinding.ActivityCreaListaBinding
+import it.cammino.risuscito.databinding.SwipeableItemBinding
+import it.cammino.risuscito.dialogs.DialogState
 import it.cammino.risuscito.dialogs.InputTextDialogFragment
 import it.cammino.risuscito.dialogs.SimpleDialogFragment
 import it.cammino.risuscito.items.SwipeableItem
 import it.cammino.risuscito.items.swipeableItem
+import it.cammino.risuscito.ui.LocaleManager.Companion.getSystemLocale
 import it.cammino.risuscito.ui.SwipeDismissTouchListener
 import it.cammino.risuscito.ui.ThemeableActivity
+import it.cammino.risuscito.utils.ThemeUtils
 import it.cammino.risuscito.viewmodels.CreaListaViewModel
 import it.cammino.risuscito.viewmodels.ViewModelWithArgumentsFactory
-import java.util.*
-import kotlin.collections.ArrayList
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 
-class CreaListaActivity : ThemeableActivity(), InputTextDialogFragment.SimpleInputCallback, SimpleDialogFragment.SimpleCallback, ItemTouchCallback, SimpleSwipeCallback.ItemSwipeCallback {
+class CreaListaActivity : ThemeableActivity(), ItemTouchCallback, SimpleSwipeCallback.ItemSwipeCallback {
 
-    private val mViewModel: CreaListaViewModel by viewModels {
+    private val mCreaListaViewModel: CreaListaViewModel by viewModels {
         ViewModelWithArgumentsFactory(application, Bundle().apply {
             putInt(ID_DA_MODIF, intent.extras?.getInt(ID_DA_MODIF, 0) ?: 0)
         })
     }
+    private val inputdialogViewModel: InputTextDialogFragment.DialogViewModel by viewModels()
+    private val simpleDialogViewModel: SimpleDialogFragment.DialogViewModel by viewModels()
 
     private var modifica: Boolean = false
     private var mAdapter: FastItemAdapter<SwipeableItem> = FastItemAdapter()
@@ -84,7 +90,6 @@ class CreaListaActivity : ThemeableActivity(), InputTextDialogFragment.SimpleInp
         binding = ActivityCreaListaBinding.inflate(layoutInflater)
         val view = binding.root
         setContentView(view)
-//        setContentView(R.layout.activity_crea_lista)
 
         modifica = intent.extras?.getBoolean(EDIT_EXISTING_LIST) == true
 
@@ -95,7 +100,7 @@ class CreaListaActivity : ThemeableActivity(), InputTextDialogFragment.SimpleInp
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
 
         val leaveBehindDrawable = IconicsDrawable(this, CommunityMaterial.Icon.cmd_delete_sweep).apply {
-            colorInt = Color.WHITE
+            colorInt = MaterialColors.getColor(this@CreaListaActivity, R.attr.colorOnPrimary, TAG)
             sizeDp = 24
             paddingDp = 2
         }
@@ -104,23 +109,23 @@ class CreaListaActivity : ThemeableActivity(), InputTextDialogFragment.SimpleInp
                 this,
                 leaveBehindDrawable,
                 ItemTouchHelper.LEFT,
-                ContextCompat.getColor(this, R.color.md_red_900))
-                .withBackgroundSwipeRight(ContextCompat.getColor(this, R.color.md_red_900))
+                MaterialColors.getColor(this, R.attr.colorPrimary, TAG))
+                .withBackgroundSwipeRight(MaterialColors.getColor(this, R.attr.colorPrimary, TAG))
                 .withLeaveBehindSwipeRight(leaveBehindDrawable)
         touchCallback.setIsDragEnabled(false)
+        touchCallback.notifyAllDrops = true
 
         mTouchHelper = ItemTouchHelper(touchCallback) // Create ItemTouchHelper and pass with parameter the SimpleDragCallback
 
-        mAdapter.onLongClickListener = { _: View?, _: IAdapter<SwipeableItem>, item: SwipeableItem, position: Int ->
+        mAdapter.addLongClickListener<SwipeableItemBinding, SwipeableItem>({ binding -> binding.swipeableText1 }) { _, position, _, item ->
             Log.d(TAG, "onItemLongClick: $position")
-            mViewModel.positionToRename = position
-            InputTextDialogFragment.Builder(
-                            this, this, RENAME)
+            mCreaListaViewModel.positionToRename = position
+            InputTextDialogFragment.show(InputTextDialogFragment.Builder(
+                    this, RENAME)
                     .title(R.string.posizione_rename)
                     .prefill(item.name?.getText(this).toString())
                     .positiveButton(R.string.aggiungi_rename)
-                    .negativeButton(R.string.cancel)
-                    .show()
+                    .negativeButton(R.string.cancel), supportFragmentManager)
             true
         }
 
@@ -146,13 +151,6 @@ class CreaListaActivity : ThemeableActivity(), InputTextDialogFragment.SimpleInp
 
         binding.textTitleDescription.requestFocus()
 
-        var iFragment = InputTextDialogFragment.findVisible(this, RENAME)
-        iFragment?.setmCallback(this)
-        iFragment = InputTextDialogFragment.findVisible(this, ADD_POSITION)
-        iFragment?.setmCallback(this)
-        val fragment = SimpleDialogFragment.findVisible(this, SAVE_LIST)
-        fragment?.setmCallback(this)
-
         binding.mainHintLayout.hintText.setText(R.string.showcase_rename_desc)
         binding.mainHintLayout.hintText.append(System.getProperty("line.separator"))
         binding.mainHintLayout.hintText.append(getString(R.string.showcase_delete_desc))
@@ -173,39 +171,106 @@ class CreaListaActivity : ThemeableActivity(), InputTextDialogFragment.SimpleInp
 
         binding.textFieldTitle.doOnTextChanged { s: CharSequence?, _: Int, _: Int, _: Int ->
             binding.collapsingToolbarLayout.title = s
-            mViewModel.tempTitle = s.toString()
+            mCreaListaViewModel.tempTitle = s.toString()
+        }
+
+        if (ThemeUtils.isDarkMode(this)) {
+            val elevatedSurfaceColor = ElevationOverlayProvider(this).compositeOverlayWithThemeSurfaceColorIfNeeded(resources.getDimension(R.dimen.design_appbar_elevation))
+            binding.collapsingToolbarLayout.setContentScrimColor(elevatedSurfaceColor)
+            binding.appBarLayout.background = ColorDrawable(elevatedSurfaceColor)
         }
 
         binding.fabCreaLista.setOnClickListener {
-            InputTextDialogFragment.Builder(
-                            this, this, ADD_POSITION)
+            InputTextDialogFragment.show(InputTextDialogFragment.Builder(
+                    this, ADD_POSITION)
                     .title(R.string.posizione_add_desc)
                     .positiveButton(R.string.aggiungi_confirm)
-                    .negativeButton(R.string.cancel)
-                    .show()
+                    .negativeButton(R.string.cancel), supportFragmentManager)
         }
 
         if (modifica)
             subscribeUiChanges()
         else {
-            if (mViewModel.tempTitle.isEmpty())
-                mViewModel.tempTitle = intent.extras?.getString(LIST_TITLE) ?: ""
-            binding.textFieldTitle.setText(mViewModel.tempTitle)
-            binding.collapsingToolbarLayout.title = mViewModel.tempTitle
-            if (mViewModel.elementi == null)
-                mViewModel.elementi = ArrayList()
-            mViewModel.elementi?.let { mAdapter.set(it) }
+            if (mCreaListaViewModel.tempTitle.isEmpty())
+                mCreaListaViewModel.tempTitle = intent.extras?.getString(LIST_TITLE) ?: ""
+            binding.textFieldTitle.setText(mCreaListaViewModel.tempTitle)
+            binding.collapsingToolbarLayout.title = mCreaListaViewModel.tempTitle
+            if (mCreaListaViewModel.elementi == null)
+                mCreaListaViewModel.elementi = ArrayList()
+            mCreaListaViewModel.elementi?.let { mAdapter.set(it) }
         }
 
         onBackPressedDispatcher.addCallback(this) {
             onBackPressedAction()
         }
 
+        inputdialogViewModel.state.observe(this) {
+            Log.d(TAG, "inputdialogViewModel state $it")
+            if (!inputdialogViewModel.handled) {
+                when (it) {
+                    is DialogState.Positive -> {
+                        when (inputdialogViewModel.mTag) {
+                            RENAME -> {
+                                inputdialogViewModel.handled = true
+                                val mElement = mAdapter.adapterItems[mCreaListaViewModel.positionToRename]
+                                mElement.setName = inputdialogViewModel.outputText
+                                mAdapter.notifyAdapterItemChanged(mCreaListaViewModel.positionToRename)
+                            }
+                            ADD_POSITION -> {
+                                inputdialogViewModel.handled = true
+                                binding.noElementsAdded.isVisible = false
+                                mAdapter.add(swipeableItem {
+                                    identifier = Utility.random(0, 5000).toLong()
+                                    touchHelper = mTouchHelper
+                                    setName = inputdialogViewModel.outputText
+                                })
+                                Log.d(TAG, "onPositive - elementi.size(): " + mAdapter.adapterItems.size)
+                                val mSharedPrefs = PreferenceManager.getDefaultSharedPreferences(this)
+                                Log.d(
+                                        TAG,
+                                        "onCreateOptionsMenu - INTRO_CREALISTA_2: " + mSharedPrefs.getBoolean(Utility.INTRO_CREALISTA_2, false))
+                                binding.mainHintLayout.mainHintLayout.isVisible = !mSharedPrefs.getBoolean(Utility.INTRO_CREALISTA_2, false)
+                            }
+                        }
+                    }
+                    is DialogState.Negative -> {
+                        inputdialogViewModel.handled = true
+                    }
+                }
+            }
+        }
+
+        simpleDialogViewModel.state.observe(this) {
+            Log.d(TAG, "simpleDialogViewModel state $it")
+            if (!simpleDialogViewModel.handled) {
+                when (it) {
+                    is DialogState.Positive -> {
+                        when (simpleDialogViewModel.mTag) {
+                            SAVE_LIST -> {
+                                simpleDialogViewModel.handled = true
+                                lifecycleScope.launch { saveList() }
+                            }
+                        }
+                    }
+                    is DialogState.Negative -> {
+                        when (simpleDialogViewModel.mTag) {
+                            SAVE_LIST -> {
+                                simpleDialogViewModel.handled = true
+                                setResult(RESULT_CANCELED)
+                                finish()
+                                Animatoo.animateSlideDown(this)
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
     }
 
     override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
-        mViewModel.elementi = mAdapter.itemAdapter.adapterItems as? ArrayList<SwipeableItem>
+        mCreaListaViewModel.elementi = mAdapter.itemAdapter.adapterItems as? ArrayList<SwipeableItem>
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
@@ -217,7 +282,7 @@ class CreaListaActivity : ThemeableActivity(), InputTextDialogFragment.SimpleInp
                 TAG,
                 "onCreateOptionsMenu - INTRO_CREALISTA: " + mSharedPrefs.getBoolean(Utility.INTRO_CREALISTA, false))
         if (!mSharedPrefs.getBoolean(Utility.INTRO_CREALISTA, false)) {
-            Handler().postDelayed(1500) {
+            Handler(Looper.getMainLooper()).postDelayed(1500) {
                 playIntro()
             }
         }
@@ -233,21 +298,21 @@ class CreaListaActivity : ThemeableActivity(), InputTextDialogFragment.SimpleInp
                 return true
             }
             R.id.action_save_list -> {
-                SaveListTask().execute(binding.textFieldTitle.text)
+                lifecycleScope.launch { saveList() }
                 return true
             }
             android.R.id.home -> {
                 if (mAdapter.adapterItems.isNotEmpty()) {
-                    SimpleDialogFragment.Builder(
-                                    this, this, SAVE_LIST)
+                    SimpleDialogFragment.show(SimpleDialogFragment.Builder(
+                            this, SAVE_LIST)
                             .title(R.string.save_list_title)
                             .content(R.string.save_list_question)
                             .positiveButton(R.string.save_exit_confirm)
-                            .negativeButton(R.string.discard_exit_confirm)
-                            .show()
+                            .negativeButton(R.string.discard_exit_confirm),
+                            supportFragmentManager)
                     return true
                 } else {
-                    setResult(Activity.RESULT_CANCELED)
+                    setResult(RESULT_CANCELED)
                     finish()
                     Animatoo.animateSlideDown(this)
                 }
@@ -257,70 +322,87 @@ class CreaListaActivity : ThemeableActivity(), InputTextDialogFragment.SimpleInp
         return false
     }
 
+    private suspend fun saveList() {
+        val mDao = RisuscitoDatabase.getInstance(this).listePersDao()
+
+        var result = 0
+        val celebrazione = ListaPersonalizzata()
+
+        if (binding.textFieldTitle.text.toString().isNotBlank()) {
+            celebrazione.name = binding.textFieldTitle.text.toString()
+        } else {
+            result += 100
+            celebrazione.name = if (modifica) withContext(lifecycleScope.coroutineContext + Dispatchers.IO) { mDao.getListById(mCreaListaViewModel.idModifica)?.titolo }
+                    ?: DEFAULT_TITLE else intent.extras?.getString(LIST_TITLE)
+                    ?: DEFAULT_TITLE
+        }
+
+        Log.d(TAG, "saveList - elementi.size(): " + mAdapter.adapterItems.size)
+        for (i in mAdapter.adapterItems.indices) {
+            mAdapter.getItem(i)?.let {
+                if (celebrazione.addPosizione(it.name?.getText(this).toString()) == -2) {
+                    Snackbar.make(
+                            binding.mainContent,
+                            R.string.lista_pers_piena,
+                            Snackbar.LENGTH_SHORT)
+                            .show()
+                    return
+                }
+                celebrazione.addCanto(it.idCanto, i)
+            }
+        }
+
+        if (celebrazione.getNomePosizione(0).equals("", ignoreCase = true)) {
+            Snackbar.make(
+                    binding.mainContent, R.string.lista_pers_vuota,
+                    Snackbar.LENGTH_SHORT)
+                    .show()
+            return
+        }
+
+        Log.d(TAG, "saveList - $celebrazione")
+
+        val listaToUpdate = ListaPers()
+        listaToUpdate.lista = celebrazione
+        listaToUpdate.titolo = celebrazione.name
+        if (modifica) {
+            listaToUpdate.id = mCreaListaViewModel.idModifica
+            withContext(lifecycleScope.coroutineContext + Dispatchers.IO) { mDao.updateLista(listaToUpdate) }
+        } else
+            withContext(lifecycleScope.coroutineContext + Dispatchers.IO) { mDao.insertLista(listaToUpdate) }
+
+        if (result == 100)
+            Toast.makeText(this, getString(R.string.no_title_edited), Toast.LENGTH_SHORT).show()
+        setResult(RESULT_OK)
+        finish()
+        Animatoo.animateSlideDown(this)
+    }
+
     private fun onBackPressedAction() {
         Log.d(TAG, "onBackPressed: ")
         if (mAdapter.adapterItems.isNotEmpty()) {
-            SimpleDialogFragment.Builder(this, this, SAVE_LIST)
+            SimpleDialogFragment.show(SimpleDialogFragment.Builder(this, SAVE_LIST)
                     .title(R.string.save_list_title)
                     .content(R.string.save_list_question)
                     .positiveButton(R.string.save_exit_confirm)
-                    .negativeButton(R.string.discard_exit_confirm)
-                    .show()
+                    .negativeButton(R.string.discard_exit_confirm),
+                    supportFragmentManager)
         } else {
-            setResult(Activity.RESULT_CANCELED)
+            setResult(RESULT_CANCELED)
             finish()
             Animatoo.animateSlideDown(this)
         }
     }
 
-    override fun onPositive(tag: String, dialog: MaterialDialog) {
-        Log.d(TAG, "onPositive: $tag")
-        when (tag) {
-            RENAME -> {
-                val mEditText = dialog.getInputField()
-                val mElement = mAdapter.adapterItems[mViewModel.positionToRename]
-                mElement.setName = mEditText.text.toString()
-                mAdapter.notifyAdapterItemChanged(mViewModel.positionToRename)
-            }
-            ADD_POSITION -> {
-                binding.noElementsAdded.isVisible = false
-                val mEditText = dialog.getInputField()
-                mAdapter.add(swipeableItem {
-                    identifier = Utility.random(0, 5000).toLong()
-                    touchHelper = mTouchHelper
-                    setName = mEditText.text.toString()
-                })
-                Log.d(TAG, "onPositive - elementi.size(): " + mAdapter.adapterItems.size)
-                val mSharedPrefs = PreferenceManager.getDefaultSharedPreferences(this)
-                Log.d(
-                        TAG,
-                        "onCreateOptionsMenu - INTRO_CREALISTA_2: " + mSharedPrefs.getBoolean(Utility.INTRO_CREALISTA_2, false))
-                binding.mainHintLayout.mainHintLayout.isVisible = !mSharedPrefs.getBoolean(Utility.INTRO_CREALISTA_2, false)
-            }
-        }
+    override fun itemTouchStartDrag(viewHolder: RecyclerView.ViewHolder) {
+        @Suppress("UNCHECKED_CAST")
+        (viewHolder as? BindingViewHolder<SwipeableItemBinding>)?.binding?.cardContainer?.isDragged = true
     }
 
-    override fun onNegative(tag: String, dialog: MaterialDialog) {
-        // no-op
-    }
-
-    override fun onPositive(tag: String) {
-        Log.d(TAG, "onPositive: $tag")
-        when (tag) {
-            SAVE_LIST ->
-                SaveListTask().execute(binding.textFieldTitle.text)
-        }
-    }
-
-    override fun onNegative(tag: String) {
-        Log.d(TAG, "onNegative: $tag")
-        when (tag) {
-            SAVE_LIST -> {
-                setResult(Activity.RESULT_CANCELED)
-                finish()
-                Animatoo.animateSlideDown(this)
-            }
-        }
+    override fun itemTouchDropped(oldPosition: Int, newPosition: Int) {
+        @Suppress("UNCHECKED_CAST")
+        val viewHolder = binding.recyclerView.findViewHolderForAdapterPosition(newPosition) as? BindingViewHolder<SwipeableItemBinding>?
+        viewHolder?.binding?.cardContainer?.isDragged = false
     }
 
     override fun itemTouchOnMove(oldPosition: Int, newPosition: Int): Boolean {
@@ -328,85 +410,62 @@ class CreaListaActivity : ThemeableActivity(), InputTextDialogFragment.SimpleInp
         return true
     }
 
-    override fun itemTouchDropped(oldPosition: Int, newPosition: Int) = Unit
-
     override fun itemSwiped(position: Int, direction: Int) {
-        // -- Option 1: Direct action --
-        // do something when swiped such as: select, remove, update, ...:
-        // A) fastItemAdapter.select(position);
-        // B) fastItemAdapter.remove(position);
-        // C) update item, set "read" if an email etc
-
-        // -- Option 2: Delayed action --
         val item = mAdapter.getItem(position) ?: return
-        item.swipedDirection = direction
 
-        val deleteHandler = Handler { message ->
-            val itemOjb = message.obj as SwipeableItem
+        mAdapter.remove(position)
+        binding.noElementsAdded.isVisible = mAdapter.adapterItemCount == 0
+        if (mAdapter.adapterItemCount == 0) binding.mainHintLayout.mainHintLayout.isVisible = false
 
-            itemOjb.swipedAction = null
-            val position12 = mAdapter.getAdapterPosition(itemOjb)
-            if (position12 != RecyclerView.NO_POSITION) {
-                mAdapter.remove(position12)
-                binding.noElementsAdded.isVisible = mAdapter.adapterItemCount == 0
-                if (mAdapter.adapterItemCount == 0) binding.mainHintLayout.mainHintLayout.isVisible = false
-            }
-            true
-        }
-
-        // This can vary depending on direction but remove & archive simulated here both results in
-        // removal from list
-        val message = Random().nextInt()
-        deleteHandler.sendMessageDelayed(Message.obtain().apply { what = message; obj = item }, 2000)
-
-        item.swipedAction = Runnable {
-            deleteHandler.removeMessages(message)
-            item.swipedDirection = 0
-            val mPosition = mAdapter.getAdapterPosition(item)
-            if (mPosition != RecyclerView.NO_POSITION)
-                mAdapter.notifyItemChanged(mPosition)
-        }
-
-        mAdapter.notifyItemChanged(position)
+        Snackbar.make(binding.mainContent, getString(R.string.generic_removed, item.name?.getText(this@CreaListaActivity)), Snackbar.LENGTH_SHORT)
+                .setAction(getString(R.string.cancel).toUpperCase(getSystemLocale(resources))) {
+                    item.swipedDirection = 0
+                    mAdapter.add(position, item)
+                    if (position != RecyclerView.NO_POSITION)
+                        mAdapter.notifyItemChanged(position)
+                    binding.noElementsAdded.isVisible = mAdapter.adapterItemCount == 0
+                    if (mAdapter.adapterItemCount == 0) binding.mainHintLayout.mainHintLayout.isVisible = false
+                }.show()
     }
 
     private fun playIntro() {
         binding.fabCreaLista.show()
+        val colorOnPrimary = MaterialColors.getColor(this, R.attr.colorOnPrimary, TAG)
         TapTargetSequence(this)
                 .continueOnCancel(true)
                 .targets(
                         TapTarget.forView(
-                                        binding.fabCreaLista,
-                                        getString(R.string.add_position),
-                                        getString(R.string.showcase_add_pos_desc))
+                                binding.fabCreaLista,
+                                getString(R.string.add_position),
+                                getString(R.string.showcase_add_pos_desc))
                                 // All options below are optional
-                                .targetCircleColorInt(Color.WHITE) // Specify a color for the target circle
+                                .targetCircleColorInt(colorOnPrimary) // Specify a color for the target circle
                                 .textTypeface(mRegularFont) // Specify a typeface for the text
-                                .titleTextColor(R.color.primary_text_default_material_dark)
-                                .textColor(R.color.secondary_text_default_material_dark)
+                                .titleTextColorInt(colorOnPrimary)
+                                .textColorInt(colorOnPrimary)
                                 .tintTarget(false)
                                 .id(1),
                         TapTarget.forToolbarMenuItem(
-                                        binding.risuscitoToolbar,
-                                        R.id.action_save_list,
-                                        getString(R.string.list_save_exit),
-                                        getString(R.string.showcase_saveexit_desc))
+                                binding.risuscitoToolbar,
+                                R.id.action_save_list,
+                                getString(R.string.list_save_exit),
+                                getString(R.string.showcase_saveexit_desc))
                                 // All options below are optional
-                                .targetCircleColorInt(Color.WHITE) // Specify a color for the target circle
+                                .targetCircleColorInt(colorOnPrimary) // Specify a color for the target circle
                                 .textTypeface(mRegularFont) // Specify a typeface for the text
-                                .titleTextColor(R.color.primary_text_default_material_dark)
-                                .textColor(R.color.secondary_text_default_material_dark)
+                                .titleTextColorInt(colorOnPrimary)
+                                .textColorInt(colorOnPrimary)
                                 .id(2),
                         TapTarget.forToolbarMenuItem(
-                                        binding.risuscitoToolbar,
-                                        R.id.action_help,
-                                        getString(R.string.showcase_end_title),
-                                        getString(R.string.showcase_help_general))
+                                binding.risuscitoToolbar,
+                                R.id.action_help,
+                                getString(R.string.showcase_end_title),
+                                getString(R.string.showcase_help_general))
                                 // All options below are optional
-                                .targetCircleColorInt(Color.WHITE) // Specify a color for the target circle
+                                .targetCircleColorInt(colorOnPrimary) // Specify a color for the target circle
                                 .textTypeface(mRegularFont) // Specify a typeface for the text
-                                .titleTextColor(R.color.primary_text_default_material_dark)
-                                .textColor(R.color.secondary_text_default_material_dark)
+                                .titleTextColorInt(colorOnPrimary)
+                                .textColorInt(colorOnPrimary)
                                 .id(3))
                 .listener(
                         object : TapTargetSequence.Listener { // The listener can listen for regular clicks, long clicks or cancels
@@ -428,19 +487,19 @@ class CreaListaActivity : ThemeableActivity(), InputTextDialogFragment.SimpleInp
     }
 
     private fun subscribeUiChanges() {
-        mViewModel.listaResult?.observe(this) { listaPers ->
+        mCreaListaViewModel.listaResult?.observe(this) { listaPers ->
 
             val celebrazione = listaPers.lista
 
-            mViewModel.elementi?.let {
+            mCreaListaViewModel.elementi?.let {
                 Log.d(TAG, "Lista già valorizzata")
                 for (elemento in it) elemento.touchHelper = mTouchHelper
             } ?: run {
                 Log.d(TAG, "Lista nulla")
-                mViewModel.elementi = ArrayList()
+                mCreaListaViewModel.elementi = ArrayList()
                 celebrazione?.let {
                     for (i in 0 until it.numPosizioni) {
-                        mViewModel.elementi?.add(
+                        mCreaListaViewModel.elementi?.add(
                                 swipeableItem {
                                     identifier = Utility.random(0, 5000).toLong()
                                     touchHelper = mTouchHelper
@@ -452,83 +511,12 @@ class CreaListaActivity : ThemeableActivity(), InputTextDialogFragment.SimpleInp
                 }
             }
 
-            mViewModel.elementi?.let { mAdapter.set(it) }
-            if (mViewModel.tempTitle.isEmpty())
-                mViewModel.tempTitle = listaPers.titolo ?: DEFAULT_TITLE
-            binding.textFieldTitle.setText(mViewModel.tempTitle)
-            binding.collapsingToolbarLayout.title = mViewModel.tempTitle
+            mCreaListaViewModel.elementi?.let { mAdapter.set(it) }
+            if (mCreaListaViewModel.tempTitle.isEmpty())
+                mCreaListaViewModel.tempTitle = listaPers.titolo ?: DEFAULT_TITLE
+            binding.textFieldTitle.setText(mCreaListaViewModel.tempTitle)
+            binding.collapsingToolbarLayout.title = mCreaListaViewModel.tempTitle
             binding.noElementsAdded.isVisible = mAdapter.adapterItemCount == 0
-        }
-    }
-
-    @SuppressLint("StaticFieldLeak")
-    private inner class SaveListTask : AsyncTask<Editable, Void, Int>() {
-
-        override fun doInBackground(vararg titleText: Editable): Int {
-
-            val mDao = RisuscitoDatabase.getInstance(this@CreaListaActivity).listePersDao()
-
-            var result = 0
-            val celebrazione = ListaPersonalizzata()
-
-            if (titleText[0].isNotBlank()) {
-                celebrazione.name = titleText[0].toString()
-            } else {
-                result += 100
-                celebrazione.name = if (modifica) mDao.getListById(mViewModel.idModifica)?.titolo
-                        ?: DEFAULT_TITLE else intent.extras?.getString(LIST_TITLE) ?: DEFAULT_TITLE
-            }
-
-            Log.d(TAG, "saveList - elementi.size(): " + mAdapter.adapterItems.size)
-            for (i in mAdapter.adapterItems.indices) {
-                mAdapter.getItem(i)?.let {
-                    if (celebrazione.addPosizione(it.name?.getText(this@CreaListaActivity).toString()) == -2) {
-                        return 1
-                    }
-                    celebrazione.addCanto(it.idCanto, i)
-                }
-            }
-
-            if (celebrazione.getNomePosizione(0).equals("", ignoreCase = true))
-                return 2
-
-            Log.d(TAG, "saveList - $celebrazione")
-
-            val listaToUpdate = ListaPers()
-            listaToUpdate.lista = celebrazione
-            listaToUpdate.titolo = celebrazione.name
-            if (modifica) {
-                listaToUpdate.id = mViewModel.idModifica
-                mDao.updateLista(listaToUpdate)
-            } else
-                mDao.insertLista(listaToUpdate)
-
-            return result
-        }
-
-        override fun onPostExecute(result: Int) {
-            super.onPostExecute(result)
-            if (result == 100)
-                Toast.makeText(this@CreaListaActivity, getString(R.string.no_title_edited), Toast.LENGTH_SHORT).show()
-            when (result) {
-                0, 100 -> {
-                    setResult(Activity.RESULT_OK)
-                    finish()
-                    Animatoo.animateSlideDown(this@CreaListaActivity)
-                }
-                1 ->
-                    Snackbar.make(
-                                    binding.mainContent,
-                                    R.string.lista_pers_piena,
-                                    Snackbar.LENGTH_SHORT)
-                            .show()
-                2 ->
-                    Snackbar.make(
-                                    binding.mainContent, R.string.lista_pers_vuota,
-                                    Snackbar.LENGTH_SHORT)
-                            .show()
-
-            }
         }
     }
 
