@@ -9,21 +9,21 @@ import android.os.SystemClock
 import android.util.Log
 import android.view.*
 import android.widget.Button
+import androidx.activity.result.ActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.core.content.edit
 import androidx.core.content.res.ResourcesCompat
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.observe
+import androidx.lifecycle.lifecycleScope
 import androidx.preference.PreferenceManager
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import androidx.viewpager2.widget.ViewPager2
-import com.afollestad.materialcab.MaterialCab.Companion.destroy
-import com.afollestad.materialdialogs.MaterialDialog
-import com.afollestad.materialdialogs.input.getInputField
 import com.blogspot.atifsoftwares.animatoolib.Animatoo
 import com.getkeepsafe.taptargetview.TapTarget
 import com.getkeepsafe.taptargetview.TapTargetSequence
+import com.google.android.material.color.MaterialColors
 import com.google.android.material.snackbar.Snackbar
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
@@ -40,15 +40,20 @@ import it.cammino.risuscito.CreaListaActivity.Companion.LIST_TITLE
 import it.cammino.risuscito.database.RisuscitoDatabase
 import it.cammino.risuscito.database.entities.ListaPers
 import it.cammino.risuscito.databinding.TabsLayout2Binding
+import it.cammino.risuscito.dialogs.DialogState
 import it.cammino.risuscito.dialogs.InputTextDialogFragment
 import it.cammino.risuscito.dialogs.SimpleDialogFragment
 import it.cammino.risuscito.ui.LocaleManager.Companion.getSystemLocale
-import it.cammino.risuscito.utils.ioThread
 import it.cammino.risuscito.viewmodels.CustomListsViewModel
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 
-class CustomLists : Fragment(), InputTextDialogFragment.SimpleInputCallback, SimpleDialogFragment.SimpleCallback {
+class CustomLists : Fragment() {
 
     private val mCustomListsViewModel: CustomListsViewModel by viewModels()
+    private val inputdialogViewModel: InputTextDialogFragment.DialogViewModel by viewModels({ requireActivity() })
+    private val simpleDialogViewModel: SimpleDialogFragment.DialogViewModel by viewModels({ requireActivity() })
     private var mSectionsPagerAdapter: SectionsPagerAdapter? = null
     private var titoliListe: Array<String?> = arrayOfNulls(0)
     private var idListe: IntArray = IntArray(0)
@@ -63,7 +68,7 @@ class CustomLists : Fragment(), InputTextDialogFragment.SimpleInputCallback, Sim
             Log.d(TAG, "mCustomListsViewModel.indexToShow: ${mCustomListsViewModel.indexToShow}")
             if (mCustomListsViewModel.indexToShow != position) {
                 mCustomListsViewModel.indexToShow = position
-                destroy()
+                mMainActivity?.actionMode?.finish()
             }
             initFabOptions(position >= 2)
         }
@@ -89,7 +94,7 @@ class CustomLists : Fragment(), InputTextDialogFragment.SimpleInputCallback, Sim
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
+        setHasOptionsMenu(true)
         mRegularFont = ResourcesCompat.getFont(requireContext(), R.font.googlesans_regular)
 
         mMainActivity = activity as? MainActivity
@@ -99,13 +104,6 @@ class CustomLists : Fragment(), InputTextDialogFragment.SimpleInputCallback, Sim
         mMainActivity?.enableFab(true)
 
         movePage = savedInstanceState != null
-
-        val iFragment = InputTextDialogFragment.findVisible(mMainActivity, NEW_LIST)
-        iFragment?.setmCallback(this)
-        var sFragment = SimpleDialogFragment.findVisible(mMainActivity, RESET_LIST)
-        sFragment?.setmCallback(this)
-        sFragment = SimpleDialogFragment.findVisible(mMainActivity, DELETE_LIST)
-        sFragment?.setmCallback(this)
 
         val mSharedPrefs = PreferenceManager.getDefaultSharedPreferences(context)
         Log.d(
@@ -134,13 +132,7 @@ class CustomLists : Fragment(), InputTextDialogFragment.SimpleInputCallback, Sim
     override fun onDestroy() {
         Log.d(TAG, "onDestroy")
         super.onDestroy()
-        destroy()
-    }
-
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
-        Log.d(TAG, "onActivityCreated")
-        setHasOptionsMenu(true)
+        mMainActivity?.actionMode?.finish()
     }
 
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
@@ -159,88 +151,12 @@ class CustomLists : Fragment(), InputTextDialogFragment.SimpleInputCallback, Sim
         return false
     }
 
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        Log.d(TAG, "onActivityResult requestCode: $requestCode")
-        super.onActivityResult(requestCode, resultCode, data)
-        if ((requestCode == TAG_CREA_LISTA || requestCode == TAG_MODIFICA_LISTA) && resultCode == Activity.RESULT_OK) {
+    private val startListEditForResult = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result: ActivityResult ->
+        if (result.resultCode == Activity.RESULT_OK) {
             Log.d(TAG, "mCustomListsViewModel.indDaModif: ${mCustomListsViewModel.indDaModif}")
             mCustomListsViewModel.indexToShow = mCustomListsViewModel.indDaModif
             movePage = true
         }
-        if (requestCode == ListaPredefinitaFragment.TAG_INSERT_PAROLA
-                || requestCode == ListaPredefinitaFragment.TAG_INSERT_EUCARESTIA
-                || requestCode == ListaPersonalizzataFragment.TAG_INSERT_PERS) {
-            Log.d(TAG, "onActivityResult resultCode: $resultCode")
-            if (resultCode == RESULT_OK || resultCode == RESULT_KO)
-                mMainActivity?.activityMainContent?.let {
-                    Snackbar.make(it, if (resultCode == RESULT_OK) R.string.list_added else R.string.present_yet, Snackbar.LENGTH_SHORT).show()
-                }
-        }
-    }
-
-    override fun onPositive(tag: String, dialog: MaterialDialog) {
-        Log.d(TAG, "onPositive: $tag")
-        when (tag) {
-            NEW_LIST -> {
-                val mEditText = dialog.getInputField()
-                mCustomListsViewModel.indDaModif = 2 + idListe.size
-                startActivityForResult(
-                        Intent(activity, CreaListaActivity::class.java).putExtras(bundleOf(LIST_TITLE to mEditText.text.toString(), EDIT_EXISTING_LIST to false)), TAG_CREA_LISTA)
-                Animatoo.animateSlideUp(activity)
-            }
-        }
-    }
-
-    override fun onNegative(tag: String, dialog: MaterialDialog) {
-        // no-op
-    }
-
-    override fun onPositive(tag: String) {
-        Log.d(TAG, "onPositive: $tag")
-        when (tag) {
-            RESET_LIST -> {
-                binding.viewPager.findViewById<Button>(R.id.button_pulisci).performClick()
-            }
-            DELETE_LIST -> {
-                binding.viewPager.currentItem = binding.viewPager.currentItem - 1
-                ioThread {
-                    val mDao = RisuscitoDatabase.getInstance(requireContext()).listePersDao()
-                    val listToDelete = ListaPers()
-                    listToDelete.id = mCustomListsViewModel.idDaCanc
-                    mDao.deleteList(listToDelete)
-                    mMainActivity?.activityMainContent?.let { mainContent ->
-                        Snackbar.make(
-                                        mainContent,
-                                        getString(R.string.list_removed)
-                                                + mCustomListsViewModel.titoloDaCanc
-                                                + "'!",
-                                        Snackbar.LENGTH_LONG)
-                                .setAction(
-                                        getString(R.string.cancel).toUpperCase(getSystemLocale(resources))
-                                ) {
-                                    if (SystemClock.elapsedRealtime() - mLastClickTime >= Utility.CLICK_DELAY) {
-                                        mLastClickTime = SystemClock.elapsedRealtime()
-                                        mCustomListsViewModel.indexToShow = mCustomListsViewModel.listaDaCanc + 2
-                                        movePage = true
-                                        ioThread {
-                                            val mListePersDao = RisuscitoDatabase.getInstance(requireContext())
-                                                    .listePersDao()
-                                            val listaToRestore = ListaPers()
-                                            listaToRestore.id = mCustomListsViewModel.idDaCanc
-                                            listaToRestore.titolo = mCustomListsViewModel.titoloDaCanc
-                                            listaToRestore.lista = mCustomListsViewModel.celebrazioneDaCanc
-                                            mListePersDao.insertLista(listaToRestore)
-                                        }
-                                    }
-                                }.show()
-                    }
-                }
-            }
-        }
-    }
-
-    override fun onNegative(tag: String) {
-        // no-op
     }
 
     private fun playIntro() {
@@ -251,29 +167,30 @@ class CustomLists : Fragment(), InputTextDialogFragment.SimpleInputCallback, Sim
             paddingDp = 4
         }
         mMainActivity?.getFab()?.let { fab ->
+            val colorOnPrimary = MaterialColors.getColor(requireContext(), R.attr.colorOnPrimary, TAG)
             TapTargetSequence(requireActivity())
                     .continueOnCancel(true)
                     .targets(
                             TapTarget.forView(
-                                            fab,
-                                            getString(R.string.showcase_listepers_title),
-                                            getString(R.string.showcase_listepers_desc1))
-                                    .targetCircleColorInt(Color.WHITE) // Specify a color for the target circle
+                                    fab,
+                                    getString(R.string.showcase_listepers_title),
+                                    getString(R.string.showcase_listepers_desc1))
+                                    .targetCircleColorInt(colorOnPrimary) // Specify a color for the target circle
                                     .textTypeface(mRegularFont) // Specify a typeface for the text
-                                    .titleTextColor(R.color.primary_text_default_material_dark)
-                                    .textColor(R.color.secondary_text_default_material_dark)
+                                    .titleTextColorInt(colorOnPrimary)
+                                    .textColorInt(colorOnPrimary)
                                     .descriptionTextSize(15)
                                     .tintTarget(false) // Whether to tint the target view's color
                             ,
                             TapTarget.forView(
-                                            fab,
-                                            getString(R.string.showcase_listepers_title),
-                                            getString(R.string.showcase_listepers_desc3))
-                                    .targetCircleColorInt(Color.WHITE) // Specify a color for the target circle
+                                    fab,
+                                    getString(R.string.showcase_listepers_title),
+                                    getString(R.string.showcase_listepers_desc3))
+                                    .targetCircleColorInt(colorOnPrimary) // Specify a color for the target circle
                                     .icon(doneDrawable)
                                     .textTypeface(mRegularFont) // Specify a typeface for the text
-                                    .titleTextColor(R.color.primary_text_default_material_dark)
-                                    .textColor(R.color.secondary_text_default_material_dark))
+                                    .titleTextColorInt(colorOnPrimary)
+                                    .textColorInt(colorOnPrimary))
                     .listener(
                             object : TapTargetSequence.Listener { // The listener can listen for regular clicks, long clicks or cancels
                                 override fun onSequenceFinish() {
@@ -310,21 +227,63 @@ class CustomLists : Fragment(), InputTextDialogFragment.SimpleInputCallback, Sim
                 movePage = false
             }
         }
-    }
 
-    private inner class SectionsPagerAdapter internal constructor(fm: Fragment) : FragmentStateAdapter(fm) {
-
-        override fun createFragment(position: Int): Fragment {
-            return when (position) {
-                0 -> ListaPredefinitaFragment.newInstance(1)
-                1 -> ListaPredefinitaFragment.newInstance(2)
-                else -> ListaPersonalizzataFragment.newInstance(idListe[position - 2])
+        inputdialogViewModel.state.observe(viewLifecycleOwner) {
+            Log.d(TAG, "inputdialogViewModel state $it")
+            if (!inputdialogViewModel.handled) {
+                when (it) {
+                    is DialogState.Positive -> {
+                        when (inputdialogViewModel.mTag) {
+                            NEW_LIST -> {
+                                inputdialogViewModel.handled = true
+                                mCustomListsViewModel.indDaModif = 2 + idListe.size
+                                startListEditForResult.launch(Intent(activity, CreaListaActivity::class.java).putExtras(bundleOf(LIST_TITLE to inputdialogViewModel.outputText, EDIT_EXISTING_LIST to false)))
+                                Animatoo.animateSlideUp(activity)
+                            }
+                        }
+                    }
+                    is DialogState.Negative -> {
+                        inputdialogViewModel.handled = true
+                    }
+                }
             }
         }
 
-        override fun getItemCount(): Int {
-            return 2 + titoliListe.size
+        simpleDialogViewModel.state.observe(viewLifecycleOwner) {
+            Log.d(TAG, "simpleDialogViewModel state $it")
+            if (!simpleDialogViewModel.handled) {
+                when (it) {
+                    is DialogState.Positive -> {
+                        when (simpleDialogViewModel.mTag) {
+                            RESET_LIST -> {
+                                simpleDialogViewModel.handled = true
+                                binding.viewPager.findViewById<Button>(R.id.button_pulisci).performClick()
+                            }
+                            DELETE_LIST -> {
+                                simpleDialogViewModel.handled = true
+                                binding.viewPager.currentItem = binding.viewPager.currentItem - 1
+                                lifecycleScope.launch { deleteList() }
+                            }
+                        }
+                    }
+                    is DialogState.Negative -> {
+                        simpleDialogViewModel.handled = true
+                    }
+                }
+            }
         }
+    }
+
+    private inner class SectionsPagerAdapter(fragment: Fragment) : FragmentStateAdapter(fragment) {
+
+        override fun createFragment(position: Int): Fragment =
+                when (position) {
+                    0 -> ListaPredefinitaFragment.newInstance(1)
+                    1 -> ListaPredefinitaFragment.newInstance(2)
+                    else -> ListaPersonalizzataFragment.newInstance(idListe[position - 2])
+                }
+
+        override fun getItemCount(): Int = 2 + titoliListe.size
 
     }
 
@@ -347,25 +306,24 @@ class CustomLists : Fragment(), InputTextDialogFragment.SimpleInputCallback, Sim
                 R.id.fab_pulisci -> {
                     mMainActivity?.let { mActivity ->
                         closeFabMenu()
-                        SimpleDialogFragment.Builder(
-                                        mActivity, this, RESET_LIST)
+                        SimpleDialogFragment.show(SimpleDialogFragment.Builder(
+                                mActivity, RESET_LIST)
                                 .title(R.string.dialog_reset_list_title)
                                 .content(R.string.reset_list_question)
                                 .positiveButton(R.string.reset_confirm)
-                                .negativeButton(R.string.cancel)
-                                .show()
+                                .negativeButton(R.string.cancel),
+                                mActivity.supportFragmentManager)
                     }
                     true
                 }
                 R.id.fab_add_lista -> {
                     mMainActivity?.let { mActivity ->
                         closeFabMenu()
-                        InputTextDialogFragment.Builder(
-                                        mActivity, this, NEW_LIST)
+                        InputTextDialogFragment.show(InputTextDialogFragment.Builder(
+                                mActivity, NEW_LIST)
                                 .title(R.string.lista_add_desc)
                                 .positiveButton(R.string.create_confirm)
-                                .negativeButton(R.string.cancel)
-                                .show()
+                                .negativeButton(R.string.cancel), mActivity.supportFragmentManager)
                     }
                     true
                 }
@@ -377,33 +335,12 @@ class CustomLists : Fragment(), InputTextDialogFragment.SimpleInputCallback, Sim
                 R.id.fab_edit_lista -> {
                     closeFabMenu()
                     mCustomListsViewModel.indDaModif = binding.viewPager.currentItem
-                    startActivityForResult(
-                            Intent(activity, CreaListaActivity::class.java).putExtras(bundleOf(ID_DA_MODIF to idListe[binding.viewPager.currentItem - 2], EDIT_EXISTING_LIST to true)),
-                            TAG_MODIFICA_LISTA)
+                    startListEditForResult.launch(Intent(activity, CreaListaActivity::class.java).putExtras(bundleOf(ID_DA_MODIF to idListe[binding.viewPager.currentItem - 2], EDIT_EXISTING_LIST to true)))
                     Animatoo.animateSlideUp(activity)
                     true
                 }
                 R.id.fab_delete_lista -> {
-                    mMainActivity?.let { mActivity ->
-                        closeFabMenu()
-                        mCustomListsViewModel.listaDaCanc = binding.viewPager.currentItem - 2
-                        mCustomListsViewModel.idDaCanc = idListe[mCustomListsViewModel.listaDaCanc]
-                        ioThread {
-                            val mDao = RisuscitoDatabase.getInstance(requireContext()).listePersDao()
-                            val lista = mDao.getListById(mCustomListsViewModel.idDaCanc)
-                            mCustomListsViewModel.titoloDaCanc = lista?.titolo
-                            mCustomListsViewModel.celebrazioneDaCanc = lista?.lista
-                            SimpleDialogFragment.Builder(
-                                            mActivity,
-                                            this,
-                                            DELETE_LIST)
-                                    .title(R.string.action_remove_list)
-                                    .content(R.string.delete_list_dialog)
-                                    .positiveButton(R.string.delete_confirm)
-                                    .negativeButton(R.string.cancel)
-                                    .show()
-                        }
-                    }
+                    lifecycleScope.launch { deleteListDialog() }
                     true
                 }
                 R.id.fab_condividi_file -> {
@@ -419,16 +356,64 @@ class CustomLists : Fragment(), InputTextDialogFragment.SimpleInputCallback, Sim
         }
 
         val click = View.OnClickListener {
-            destroy()
+            mMainActivity?.actionMode?.finish()
             toggleFabMenu()
         }
 
         mMainActivity?.initFab(true, icon, click, actionListener, customList)
     }
 
+    private suspend fun deleteListDialog() {
+        mMainActivity?.let { mActivity ->
+            closeFabMenu()
+            mCustomListsViewModel.listaDaCanc = binding.viewPager.currentItem - 2
+            mCustomListsViewModel.idDaCanc = idListe[mCustomListsViewModel.listaDaCanc]
+            val mDao = RisuscitoDatabase.getInstance(requireContext()).listePersDao()
+            val lista = withContext(lifecycleScope.coroutineContext + Dispatchers.IO) { mDao.getListById(mCustomListsViewModel.idDaCanc) }
+            mCustomListsViewModel.titoloDaCanc = lista?.titolo
+            mCustomListsViewModel.celebrazioneDaCanc = lista?.lista
+            SimpleDialogFragment.show(SimpleDialogFragment.Builder(
+                    mActivity,
+                    DELETE_LIST)
+                    .title(R.string.action_remove_list)
+                    .content(R.string.delete_list_dialog)
+                    .positiveButton(R.string.delete_confirm)
+                    .negativeButton(R.string.cancel),
+                    mActivity.supportFragmentManager)
+        }
+    }
+
+    private suspend fun deleteList() {
+        val mDao = RisuscitoDatabase.getInstance(requireContext()).listePersDao()
+        val listToDelete = ListaPers()
+        listToDelete.id = mCustomListsViewModel.idDaCanc
+        withContext(lifecycleScope.coroutineContext + Dispatchers.IO) { mDao.deleteList(listToDelete) }
+        mMainActivity?.activityMainContent?.let { mainContent ->
+            Snackbar.make(
+                    mainContent,
+                    getString(R.string.list_removed)
+                            + mCustomListsViewModel.titoloDaCanc
+                            + "'!",
+                    Snackbar.LENGTH_LONG)
+                    .setAction(
+                            getString(R.string.cancel).toUpperCase(getSystemLocale(resources))
+                    ) {
+                        if (SystemClock.elapsedRealtime() - mLastClickTime >= Utility.CLICK_DELAY) {
+                            mLastClickTime = SystemClock.elapsedRealtime()
+                            mCustomListsViewModel.indexToShow = mCustomListsViewModel.listaDaCanc + 2
+                            movePage = true
+                            val mListePersDao = RisuscitoDatabase.getInstance(requireContext()).listePersDao()
+                            val listaToRestore = ListaPers()
+                            listaToRestore.id = mCustomListsViewModel.idDaCanc
+                            listaToRestore.titolo = mCustomListsViewModel.titoloDaCanc
+                            listaToRestore.lista = mCustomListsViewModel.celebrazioneDaCanc
+                            lifecycleScope.launch(Dispatchers.IO) { mListePersDao.insertLista(listaToRestore) }
+                        }
+                    }.show()
+        }
+    }
+
     companion object {
-        const val TAG_CREA_LISTA = 111
-        const val TAG_MODIFICA_LISTA = 222
         const val RESULT_OK = 0
         const val RESULT_KO = -1
         const val RESULT_CANCELED = -2
