@@ -1,60 +1,62 @@
 package it.cammino.risuscito
 
-import android.app.Activity
 import android.graphics.Color
-import android.os.AsyncTask
+import android.os.Build
 import android.os.Bundle
 import android.os.Handler
+import android.os.Looper
+import android.util.Base64
 import android.util.Log
-import android.webkit.WebView
-import android.webkit.WebViewClient
+import android.view.View
 import androidx.activity.addCallback
-import androidx.core.view.postDelayed
+import androidx.lifecycle.lifecycleScope
 import com.blogspot.atifsoftwares.animatoolib.Animatoo
 import com.mikepenz.iconics.IconicsDrawable
 import com.mikepenz.iconics.typeface.library.community.material.CommunityMaterial
 import com.mikepenz.iconics.utils.colorInt
 import com.mikepenz.iconics.utils.paddingDp
 import com.mikepenz.iconics.utils.sizeDp
-import io.multifunctions.letCheckNull
 import it.cammino.risuscito.database.RisuscitoDatabase
 import it.cammino.risuscito.database.entities.Canto
 import it.cammino.risuscito.databinding.ActivityPaginaRenderFullscreenBinding
+import it.cammino.risuscito.ui.InitialScrollWebClient
 import it.cammino.risuscito.ui.ThemeableActivity
-import java.lang.ref.WeakReference
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Runnable
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
+import java.nio.charset.Charset
 
 class PaginaRenderFullScreen : ThemeableActivity() {
+
     private var currentCanto: Canto? = null
     var speedValue: Int = 0
     private var scrollPlaying: Boolean = false
     var idCanto: Int = 0
-    private var urlCanto: String? = null
-    private val mHandler = Handler()
+    private lateinit var htmlContent: String
+    private val mHandler = Handler(Looper.getMainLooper())
     private val mScrollDown: Runnable = object : Runnable {
         override fun run() {
             try {
-                findViewById<WebView>(R.id.canto_view).scrollBy(0, speedValue)
+                binding.cantoView.scrollBy(0, speedValue)
             } catch (e: NumberFormatException) {
-                findViewById<WebView>(R.id.canto_view).scrollBy(0, 0)
+                binding.cantoView.scrollBy(0, 0)
             }
 
             mHandler.postDelayed(this, 700)
         }
     }
-    private var mLUtils: LUtils? = null
-
     private lateinit var binding: ActivityPaginaRenderFullscreenBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        mLUtils = LUtils.getInstance(this)
-        mLUtils?.goFullscreen()
         super.onCreate(savedInstanceState)
+        mViewModel.mLUtils.goFullscreen()
         binding = ActivityPaginaRenderFullscreenBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
         // recupera il numero della pagina da visualizzare dal parametro passato dalla chiamata
         val bundle = this.intent.extras
-        urlCanto = bundle?.getString(Utility.URL_CANTO)
+        htmlContent = bundle?.getString(Utility.HTML_CONTENT) ?: ""
         speedValue = bundle?.getInt(Utility.SPEED_VALUE) ?: 0
         scrollPlaying = bundle?.getBoolean(Utility.SCROLL_PLAYING) ?: false
         idCanto = bundle?.getInt(Utility.ID_CANTO) ?: 0
@@ -65,26 +67,49 @@ class PaginaRenderFullScreen : ThemeableActivity() {
             paddingDp = 2
         }
         binding.fabFullscreenOff.setImageDrawable(icon)
-        binding.fabFullscreenOff.setOnClickListener { saveZoom() }
+        binding.fabFullscreenOff.setOnClickListener { lifecycleScope.launch { saveZoom() } }
 
         onBackPressedDispatcher.addCallback(this) {
             onBackPressedAction()
         }
+
+        lifecycleScope.launch { loadWebView() }
+
     }
 
     private fun onBackPressedAction() {
         Log.d(TAG, "onBackPressed: ")
-        saveZoom()
+        lifecycleScope.launch { saveZoom() }
     }
 
-    public override fun onResume() {
-        super.onResume()
+    private suspend fun saveZoom() {
+        @Suppress("DEPRECATION")
+        currentCanto?.let {
+            it.zoom = (binding.cantoView.scale * 100).toInt()
+            it.scrollX = binding.cantoView.scrollX
+            it.scrollY = binding.cantoView.scrollY
+            Log.d(TAG, "it.id ${it.id} / it.zoom ${it.zoom} / it.scrollX ${it.scrollX} / it.scrollY ${it.scrollY}")
+            withContext(lifecycleScope.coroutineContext + Dispatchers.IO) {
+                RisuscitoDatabase.getInstance(applicationContext).cantoDao().updateCanto(it)
+            }
+        }
+        finish()
+        Animatoo.animateZoom(this)
+    }
 
-        findViewById<WebView>(R.id.canto_view).loadUrl(urlCanto)
-        if (scrollPlaying)
-            mScrollDown.run()
+    private suspend fun loadWebView() {
+        val mDao = RisuscitoDatabase.getInstance(this).cantoDao()
+        currentCanto = withContext(lifecycleScope.coroutineContext + Dispatchers.IO) {
+            mDao.getCantoById(idCanto)
+        }
 
-        val webSettings = findViewById<WebView>(R.id.canto_view).settings
+        // fix per crash su android 4.1
+        if (Build.VERSION.SDK_INT == Build.VERSION_CODES.JELLY_BEAN)
+            binding.cantoView.setLayerType(View.LAYER_TYPE_SOFTWARE, null)
+
+        loadContentIntoWebView(htmlContent)
+
+        val webSettings = binding.cantoView.settings
         webSettings.useWideViewPort = true
         webSettings.setSupportZoom(true)
         webSettings.loadWithOverviewMode = true
@@ -92,80 +117,40 @@ class PaginaRenderFullScreen : ThemeableActivity() {
         webSettings.builtInZoomControls = true
         webSettings.displayZoomControls = false
 
-        findViewById<WebView>(R.id.canto_view).webViewClient = MyWebViewClient()
-    }
-
-    private fun saveZoom() {
-        @Suppress("DEPRECATION")
-        //aggiunto per evitare che la pagina venga chiusa troppo velocemente prima del caricamento del canto
         currentCanto?.let {
-            it.zoom = (findViewById<WebView>(R.id.canto_view).scale * 100).toInt()
-            it.scrollX = findViewById<WebView>(R.id.canto_view).scrollX
-            it.scrollY = findViewById<WebView>(R.id.canto_view).scrollY
-            Log.d(TAG, "it.id ${it.id} / it.zoom ${it.zoom} / it.scrollX ${it.scrollX} / it.scrollY ${it.scrollY}")
-            ZoomSaverTask(this, it).execute()
-            return
+            if (it.zoom > 0) binding.cantoView.setInitialScale(it.zoom)
         }
-        finish()
-        Animatoo.animateZoom(this)
+
+        binding.cantoView.webViewClient = InitialScrollWebClient(currentCanto)
+
+        if (scrollPlaying)
+            mScrollDown.run()
+
     }
 
-    private inner class MyWebViewClient : WebViewClient() {
-        override fun onPageFinished(view: WebView, url: String) {
-            view.postDelayed(500) {
-                ZoomLoaderTask(this@PaginaRenderFullScreen).execute()
-            }
-            super.onPageFinished(view, url)
-        }
+//    private inner class MyWebViewClient : WebViewClient() {
+//        override fun onPageFinished(view: WebView, url: String) {
+//            view.postDelayed(600) {
+//                if ((currentCanto?.scrollX
+//                                ?: 0) > 0 || (currentCanto?.scrollY ?: 0) > 0)
+//                    view.scrollTo(
+//                            currentCanto?.scrollX
+//                                    ?: 0, currentCanto?.scrollY ?: 0)
+//            }
+//            super.onPageFinished(view, url)
+//        }
+//    }
+
+    private fun loadContentIntoWebView(content: String?) {
+        if (!content.isNullOrEmpty()) binding.cantoView.loadData(Base64.encodeToString(
+                content.toByteArray(Charset.forName(ECONDING_UTF8)),
+                Base64.DEFAULT), DEFAULT_MIME_TYPE, ECONDING_BASE64)
     }
 
     companion object {
         private val TAG = PaginaRenderFullScreen::class.java.canonicalName
-
-        private class ZoomSaverTask internal constructor(activity: Activity, private val canto: Canto) : AsyncTask<Void, Void, Int>() {
-
-            private val activityReference: WeakReference<Activity> = WeakReference(activity)
-
-            override fun doInBackground(vararg params: Void): Int? {
-                activityReference.get()?.let {
-                    val mDao = RisuscitoDatabase.getInstance(it).cantoDao()
-                    Log.d(TAG, "canto.id ${canto.id} / canto.zoom ${canto.zoom} / canto.scrollX ${canto.scrollX} / canto.scrollY ${canto.scrollY}")
-                    mDao.updateCanto(canto)
-                }
-                return 0
-            }
-
-            override fun onPostExecute(integer: Int?) {
-                super.onPostExecute(integer)
-                activityReference.get()?.let {
-                    it.finish()
-                    Animatoo.animateZoom(it)
-                }
-            }
-        }
-
-        private class ZoomLoaderTask internal constructor(activity: PaginaRenderFullScreen) : AsyncTask<Void, Void, Int>() {
-
-            private val activityReference: WeakReference<PaginaRenderFullScreen> = WeakReference(activity)
-
-            override fun doInBackground(vararg params: Void): Int? {
-                activityReference.get()?.let {
-                    val mDao = RisuscitoDatabase.getInstance(it).cantoDao()
-                    it.currentCanto = mDao.getCantoById(it.idCanto)
-                }
-                return 0
-            }
-
-            override fun onPostExecute(integer: Int?) {
-                super.onPostExecute(integer)
-                val apiResult = Pair(activityReference.get(), activityReference.get()?.currentCanto)
-                apiResult.letCheckNull { activity, canto ->
-                    Log.d(TAG, "onPostExecute: ${canto.zoom} - ${canto.scrollX} - ${canto.scrollY}")
-                    if (canto.zoom > 0) activity.findViewById<WebView>(R.id.canto_view).setInitialScale(canto.zoom)
-                    if (canto.scrollX > 0 || canto.scrollY > 0)
-                        activity.findViewById<WebView>(R.id.canto_view).scrollTo(canto.scrollX, canto.scrollY)
-                }
-            }
-        }
+        private const val DEFAULT_MIME_TYPE = "text/html; charset=utf-8"
+        private const val ECONDING_UTF8 = "utf-8"
+        private const val ECONDING_BASE64 = "base64"
     }
 }
