@@ -1,10 +1,6 @@
 package it.cammino.risuscito
 
 import android.annotation.TargetApi
-import android.content.BroadcastReceiver
-import android.content.Context
-import android.content.Intent
-import android.content.IntentFilter
 import android.os.Build
 import android.os.Bundle
 import android.util.Log
@@ -16,35 +12,23 @@ import androidx.core.view.GravityCompat.START
 import androidx.core.view.isInvisible
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
-import androidx.localbroadcastmanager.content.LocalBroadcastManager
+import androidx.fragment.app.viewModels
 import androidx.preference.PreferenceManager
 import com.google.android.gms.common.SignInButton
 import com.google.android.material.snackbar.Snackbar
 import com.michaelflisar.changelog.ChangelogBuilder
 import it.cammino.risuscito.databinding.ActivityRisuscitoBinding
 import it.cammino.risuscito.utils.ThemeUtils
+import it.cammino.risuscito.viewmodels.MainActivityViewModel
 import pub.devrel.easypermissions.AfterPermissionGranted
 import pub.devrel.easypermissions.EasyPermissions
 import pub.devrel.easypermissions.PermissionRequest
 
 class Risuscito : Fragment(), EasyPermissions.PermissionCallbacks {
 
-    private var mMainActivity: MainActivity? = null
-    private val signInVisibility = object : BroadcastReceiver() {
-        override fun onReceive(context: Context, intent: Intent) {
-            // Implement UI change code here once notification is received
-            try {
-                Log.d(javaClass.name, "BROADCAST_SIGNIN_VISIBLE")
-                Log.d(
-                        javaClass.name,
-                        "DATA_VISIBLE: " + intent.getBooleanExtra(DATA_VISIBLE, false))
-                binding.signInButton.isVisible = intent.getBooleanExtra(DATA_VISIBLE, false)
-            } catch (e: IllegalArgumentException) {
-                Log.e(javaClass.name, e.localizedMessage, e)
-            }
+    private val activityViewModel: MainActivityViewModel by viewModels({ requireActivity() })
 
-        }
-    }
+    private var mMainActivity: MainActivity? = null
 
     private var _binding: ActivityRisuscitoBinding? = null
 
@@ -74,16 +58,23 @@ class Risuscito : Fragment(), EasyPermissions.PermissionCallbacks {
         checkStoragePermissions()
 
         binding.imageView1.setOnClickListener {
-            if (mMainActivity?.isOnTablet == false)
+            if (!activityViewModel.isOnTablet)
                 mMainActivity?.activityDrawer?.openDrawer(START)
         }
 
         binding.signInButton.setSize(SignInButton.SIZE_WIDE)
         binding.signInButton.isInvisible = PreferenceManager.getDefaultSharedPreferences(context).getBoolean(Utility.SIGNED_IN, false)
         binding.signInButton.setOnClickListener {
-            mMainActivity?.setShowSnackbar()
+            PreferenceManager.getDefaultSharedPreferences(context).edit { putBoolean(Utility.SIGN_IN_REQUESTED, true) }
+            activityViewModel.showSnackbar = true
             mMainActivity?.signIn()
         }
+
+        activityViewModel.signedIn.observe(viewLifecycleOwner) {
+            binding.signInButton.isVisible = !it
+        }
+
+        setHasOptionsMenu(true)
 
         Log.d(TAG, "getVersionCodeWrapper(): ${getVersionCodeWrapper()}")
 
@@ -94,21 +85,7 @@ class Risuscito : Fragment(), EasyPermissions.PermissionCallbacks {
                 .withTitle(getString(R.string.dialog_change_title)) // provide a custom title if desired, default one is "Changelog <VERSION>"
                 .withOkButtonLabel(getString(R.string.ok)) // provide a custom ok button text if desired, default one is "OK"
                 .buildAndShowDialog(mMainActivity, ThemeUtils.isDarkMode(requireContext())) // second parameter defines, if the dialog has a dark or light theme
-    }
 
-    override fun onResume() {
-        super.onResume()
-        LocalBroadcastManager.getInstance(requireContext()).registerReceiver(signInVisibility, IntentFilter(BROADCAST_SIGNIN_VISIBLE))
-    }
-
-    override fun onPause() {
-        super.onPause()
-        LocalBroadcastManager.getInstance(requireContext()).unregisterReceiver(signInVisibility)
-    }
-
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
-        setHasOptionsMenu(true)
     }
 
     override fun onRequestPermissionsResult(
@@ -125,9 +102,9 @@ class Risuscito : Fragment(), EasyPermissions.PermissionCallbacks {
                         requireContext(), android.Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
             EasyPermissions.requestPermissions(
                     PermissionRequest.Builder(
-                                    this,
-                                    Utility.WRITE_STORAGE_RC,
-                                    android.Manifest.permission.WRITE_EXTERNAL_STORAGE)
+                            this,
+                            Utility.WRITE_STORAGE_RC,
+                            android.Manifest.permission.WRITE_EXTERNAL_STORAGE)
                             .setRationale(R.string.external_storage_pref_rationale)
                             .build())
         }
@@ -175,7 +152,5 @@ class Risuscito : Fragment(), EasyPermissions.PermissionCallbacks {
 
     companion object {
         private val TAG = Risuscito::class.java.canonicalName
-        const val BROADCAST_SIGNIN_VISIBLE = "it.cammino.risuscito.signin.SIGNIN_VISIBLE"
-        const val DATA_VISIBLE = "it.cammino.risuscito.signin.data.DATA_VISIBLE"
     }
 }
