@@ -1,8 +1,6 @@
 package it.cammino.risuscito
 
 import android.content.Intent
-import android.os.AsyncTask
-import android.os.AsyncTask.Status
 import android.os.Bundle
 import android.os.SystemClock
 import android.util.Log
@@ -17,41 +15,48 @@ import androidx.core.view.isVisible
 import androidx.core.widget.doOnTextChanged
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.observe
+import androidx.lifecycle.lifecycleScope
 import androidx.preference.PreferenceManager
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.crashlytics.android.Crashlytics
+import com.google.firebase.crashlytics.FirebaseCrashlytics
 import com.mikepenz.fastadapter.IAdapter
 import com.mikepenz.fastadapter.adapters.FastItemAdapter
 import it.cammino.risuscito.database.RisuscitoDatabase
 import it.cammino.risuscito.database.entities.ListaPers
 import it.cammino.risuscito.databinding.SearchLayoutBinding
+import it.cammino.risuscito.dialogs.DialogState
 import it.cammino.risuscito.dialogs.SimpleDialogFragment
 import it.cammino.risuscito.items.SimpleItem
 import it.cammino.risuscito.ui.LocaleManager.Companion.getSystemLocale
 import it.cammino.risuscito.utils.ListeUtils
-import it.cammino.risuscito.utils.ioThread
+import it.cammino.risuscito.viewmodels.MainActivityViewModel
 import it.cammino.risuscito.viewmodels.SimpleIndexViewModel
 import it.cammino.risuscito.viewmodels.ViewModelWithArgumentsFactory
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.isActive
+import kotlinx.coroutines.launch
 import org.xmlpull.v1.XmlPullParserException
 import java.io.IOException
 import java.io.InputStream
-import java.lang.ref.WeakReference
+import java.text.Collator
 
-class SearchFragment : Fragment(), SimpleDialogFragment.SimpleCallback {
+class SearchFragment : Fragment() {
 
     private val mViewModel: SimpleIndexViewModel by viewModels {
         ViewModelWithArgumentsFactory(requireActivity().application, Bundle().apply { putInt(Utility.TIPO_LISTA, 0) })
     }
+    private val simpleDialogViewModel: SimpleDialogFragment.DialogViewModel by viewModels({ requireActivity() })
+    private val activityViewModel: MainActivityViewModel by viewModels({ requireActivity() })
 
-    internal val cantoAdapter: FastItemAdapter<SimpleItem> = FastItemAdapter()
-    private lateinit var aTexts: Array<Array<String?>>
+    private var job: Job = Job()
+
+    private val cantoAdapter: FastItemAdapter<SimpleItem> = FastItemAdapter()
 
     private var listePersonalizzate: List<ListaPers>? = null
-    private var mLUtils: LUtils? = null
-    private var searchTask: SearchTask? = null
+
     private var mLastClickTime: Long = 0
     private var mMainActivity: MainActivity? = null
     private lateinit var mPopupMenu: PopupMenu
@@ -89,24 +94,17 @@ class SearchFragment : Fragment(), SimpleDialogFragment.SimpleCallback {
 
         try {
             val inputStream: InputStream = resources.openRawResource(R.raw.fileout)
-            aTexts = CantiXmlParser().parse(inputStream)
+            mViewModel.aTexts = CantiXmlParser().parse(inputStream)
             inputStream.close()
         } catch (e: XmlPullParserException) {
             Log.e(TAG, "Error:", e)
-            Crashlytics.logException(e)
+            FirebaseCrashlytics.getInstance().recordException(e)
         } catch (e: IOException) {
             Log.e(TAG, "Error:", e)
-            Crashlytics.logException(e)
+            FirebaseCrashlytics.getInstance().recordException(e)
         }
 
-        mLUtils = LUtils.getInstance(requireActivity())
-
-        var sFragment = SimpleDialogFragment.findVisible(mMainActivity, SEARCH_REPLACE)
-        sFragment?.setmCallback(this)
-        sFragment = SimpleDialogFragment.findVisible(mMainActivity, SEARCH_REPLACE_2)
-        sFragment?.setmCallback(this)
-
-        ioThread { listePersonalizzate = RisuscitoDatabase.getInstance(requireContext()).listePersDao().all }
+        lifecycleScope.launch(Dispatchers.IO) { listePersonalizzate = RisuscitoDatabase.getInstance(requireContext()).listePersDao().all }
 
         subscribeUiCanti()
 
@@ -118,7 +116,7 @@ class SearchFragment : Fragment(), SimpleDialogFragment.SimpleCallback {
                 mLastClickTime = SystemClock.elapsedRealtime()
                 val intent = Intent(requireActivity().applicationContext, PaginaRenderActivity::class.java)
                 intent.putExtras(bundleOf(Utility.PAGINA to item.source?.getText(requireContext()), Utility.ID_CANTO to item.id))
-                mLUtils?.startActivityWithTransition(intent)
+                activityViewModel.mLUtils.startActivityWithTransition(intent)
                 consume = true
             }
             consume
@@ -133,8 +131,8 @@ class SearchFragment : Fragment(), SimpleDialogFragment.SimpleCallback {
         cantoAdapter.setHasStableIds(true)
 
         binding.matchedList.adapter = cantoAdapter
-        val llm = if (mMainActivity?.isGridLayout == true)
-            GridLayoutManager(context, if (mMainActivity?.hasThreeColumns == true) 3 else 2)
+        val llm = if (activityViewModel.isGridLayout)
+            GridLayoutManager(context, if (activityViewModel.hasThreeColumns) 3 else 2)
         else
             LinearLayoutManager(context)
         binding.matchedList.layoutManager = llm
@@ -153,6 +151,7 @@ class SearchFragment : Fragment(), SimpleDialogFragment.SimpleCallback {
         }
 
         binding.textFieldRicerca.doOnTextChanged { s: CharSequence?, _: Int, _: Int, _: Int ->
+            job.cancel()
             ricercaStringa(s.toString())
         }
 
@@ -164,92 +163,43 @@ class SearchFragment : Fragment(), SimpleDialogFragment.SimpleCallback {
             it.isChecked = !it.isChecked
             mViewModel.advancedSearch = it.isChecked
             binding.textBoxRicerca.hint = if (mViewModel.advancedSearch) getString(R.string.advanced_search_subtitle) else getString(R.string.fast_search_subtitle)
+            job.cancel()
             ricercaStringa(binding.textFieldRicerca.text.toString())
             true
         }
 
         binding.moreOptions.setOnClickListener {
+            mPopupMenu.menu.findItem(R.id.advanced_search).isChecked = mViewModel.advancedSearch
             mPopupMenu.show()
         }
 
     }
 
-    override fun onDestroy() {
-        Log.d(TAG, "onDestroy")
-        searchTask?.let {
-            if (it.status == Status.RUNNING) it.cancel(true)
-        }
-        super.onDestroy()
-    }
-
-    override fun onPositive(tag: String) {
-        Log.d(TAG, "onPositive: $tag")
-        when (tag) {
-            SEARCH_REPLACE -> {
-                listePersonalizzate?.let {
-                    it[mViewModel.idListaClick]
-                            .lista?.addCanto(mViewModel.idDaAgg.toString(), mViewModel.idPosizioneClick)
-                    ListeUtils.updateListaPersonalizzata(this, it[mViewModel.idListaClick])
-                }
-            }
-            SEARCH_REPLACE_2 ->
-                ListeUtils.updatePosizione(this, mViewModel.idDaAgg, mViewModel.idListaDaAgg, mViewModel.posizioneDaAgg)
-        }
-    }
-
-    override fun onNegative(tag: String) {
-        // no-op
-    }
-
     private fun ricercaStringa(s: String) {
-        // abilita il pulsante solo se la stringa ha più di 3 caratteri, senza contare gli spazi
-        if (s.trim { it <= ' ' }.length >= 3) {
-            searchTask?.let {
-                if (it.status == Status.RUNNING) it.cancel(true)
-            }
-            searchTask = SearchTask(this)
-            searchTask?.execute(binding.textFieldRicerca.text.toString(), mViewModel.advancedSearch)
-        } else {
-            if (s.isEmpty()) {
-                searchTask?.let {
-                    if (it.status == Status.RUNNING) it.cancel(true)
-                }
+        job = lifecycleScope.launch {
+            // abilita il pulsante solo se la stringa ha più di 3 caratteri, senza contare gli spazi
+            if (s.trim { it <= ' ' }.length >= 3) {
                 binding.searchNoResults.isVisible = false
-                binding.matchedList.isVisible = false
-                cantoAdapter.clear()
-                binding.searchProgress.isVisible = false
-            }
-        }
-    }
+                binding.searchProgress.isVisible = true
+                val titoliResult = ArrayList<SimpleItem>()
 
-    private class SearchTask internal constructor(fragment: SearchFragment) : AsyncTask<Any, Void, ArrayList<SimpleItem>>() {
-
-        private val fragmentReference: WeakReference<SearchFragment> = WeakReference(fragment)
-
-        override fun doInBackground(vararg sSearchText: Any): ArrayList<SimpleItem> {
-
-            val titoliResult = ArrayList<SimpleItem>()
-
-            Log.d(TAG, "STRINGA: ${sSearchText[0]}")
-            Log.d(TAG, "ADVANCED: ${sSearchText[1]}")
-            val s = sSearchText[0] as? String ?: ""
-            val advanced = sSearchText[1] as? Boolean ?: false
-            fragmentReference.get()?.let { fragment ->
-                if (advanced) {
+                Log.d(TAG, "performSearch STRINGA: $s")
+                Log.d(TAG, "performSearch ADVANCED: ${mViewModel.advancedSearch}")
+                if (mViewModel.advancedSearch) {
                     val words = s.split("\\W".toRegex()).dropLastWhile { it.isEmpty() }.toTypedArray()
 
-                    for (aText in fragment.aTexts) {
-                        if (isCancelled) return titoliResult
+                    for (aText in mViewModel.aTexts) {
+                        if (!isActive) return@launch
 
                         if (aText[0] == null || aText[0].equals("", ignoreCase = true)) break
 
                         var found = true
                         for (word in words) {
-                            if (isCancelled) return titoliResult
+                            if (!isActive) return@launch
 
                             if (word.trim { it <= ' ' }.length > 1) {
                                 var text = word.trim { it <= ' ' }
-                                text = text.toLowerCase(getSystemLocale(fragment.resources))
+                                text = text.toLowerCase(getSystemLocale(resources))
                                 text = Utility.removeAccents(text)
 
                                 if (aText[1]?.contains(text) != true) found = false
@@ -257,53 +207,77 @@ class SearchFragment : Fragment(), SimpleDialogFragment.SimpleCallback {
                         }
 
                         if (found) {
-                            Log.d(TAG, "aText[0]: ${aText[0]}")
-                            fragment.mViewModel.titoli.sortedBy { it.title?.getText(fragment.requireContext()) }
+                            Log.d(tag, "aText[0]: ${aText[0]}")
+                            mViewModel.titoli
                                     .filter { (aText[0] ?: "") == it.undecodedSource }
                                     .forEach {
-                                        if (isCancelled) return titoliResult
+                                        if (!isActive) return@launch
                                         titoliResult.add(it.apply { filter = "" })
                                     }
                         }
                     }
                 } else {
-                    val stringa = Utility.removeAccents(s).toLowerCase(getSystemLocale(fragment.resources))
-                    Log.d(TAG, "onTextChanged: stringa $stringa")
-                    fragment.mViewModel.titoli.sortedBy { it.title?.getText(fragment.requireContext()) }
+                    val stringa = Utility.removeAccents(s).toLowerCase(getSystemLocale(resources))
+                    Log.d(tag, "performSearch onTextChanged: stringa $stringa")
+                    mViewModel.titoli
                             .filter {
-                                Utility.removeAccents(it.title?.getText(fragment.requireContext())
-                                        ?: "").toLowerCase(getSystemLocale(fragment.resources)).contains(stringa)
+                                Utility.removeAccents(it.title?.getText(requireContext())
+                                        ?: "").toLowerCase(getSystemLocale(resources)).contains(stringa)
                             }
                             .forEach {
-                                if (isCancelled) return titoliResult
+                                if (!isActive) return@launch
                                 titoliResult.add(it.apply { filter = stringa })
                             }
                 }
+                if (isActive) {
+                    cantoAdapter.set(titoliResult.sortedWith(compareBy(Collator.getInstance(getSystemLocale(resources))) { it.title?.getText(requireContext()) }))
+                    binding.searchProgress.isVisible = false
+                    binding.searchNoResults.isVisible = cantoAdapter.adapterItemCount == 0
+                    binding.matchedList.isGone = cantoAdapter.adapterItemCount == 0
+                }
+            } else {
+                if (s.isEmpty()) {
+                    binding.searchNoResults.isVisible = false
+                    binding.matchedList.isVisible = false
+                    cantoAdapter.clear()
+                    binding.searchProgress.isVisible = false
+                }
             }
-            return titoliResult
-        }
-
-        override fun onPreExecute() {
-            super.onPreExecute()
-            if (isCancelled) return
-            fragmentReference.get()?.binding?.searchNoResults?.isVisible = false
-            fragmentReference.get()?.binding?.searchProgress?.isVisible = true
-        }
-
-        override fun onPostExecute(titoliResult: ArrayList<SimpleItem>) {
-            super.onPostExecute(titoliResult)
-            if (isCancelled) return
-            fragmentReference.get()?.cantoAdapter?.set(titoliResult)
-            fragmentReference.get()?.binding?.searchProgress?.isVisible = false
-            fragmentReference.get()?.binding?.searchNoResults?.isVisible = fragmentReference.get()?.cantoAdapter?.adapterItemCount == 0
-            fragmentReference.get()?.binding?.matchedList?.isGone = fragmentReference.get()?.cantoAdapter?.adapterItemCount == 0
         }
     }
 
     private fun subscribeUiCanti() {
         mViewModel.itemsResult?.observe(viewLifecycleOwner) { canti ->
-            mViewModel.titoli = canti.sortedBy { it.title?.getText(requireContext()) }
+            mViewModel.titoli = canti.sortedWith(compareBy(Collator.getInstance(getSystemLocale(resources))) { it.title?.getText(requireContext()) })
         }
+
+        simpleDialogViewModel.state.observe(viewLifecycleOwner) {
+            Log.d(TAG, "simpleDialogViewModel state $it")
+            if (!simpleDialogViewModel.handled) {
+                when (it) {
+                    is DialogState.Positive -> {
+                        when (simpleDialogViewModel.mTag) {
+                            SEARCH_REPLACE -> {
+                                simpleDialogViewModel.handled = true
+                                listePersonalizzate?.let { lista ->
+                                    lista[mViewModel.idListaClick]
+                                            .lista?.addCanto(mViewModel.idDaAgg.toString(), mViewModel.idPosizioneClick)
+                                    ListeUtils.updateListaPersonalizzata(this, lista[mViewModel.idListaClick])
+                                }
+                            }
+                            SEARCH_REPLACE_2 -> {
+                                simpleDialogViewModel.handled = true
+                                ListeUtils.updatePosizione(this, mViewModel.idDaAgg, mViewModel.idListaDaAgg, mViewModel.posizioneDaAgg)
+                            }
+                        }
+                    }
+                    is DialogState.Negative -> {
+                        simpleDialogViewModel.handled = true
+                    }
+                }
+            }
+        }
+
     }
 
     companion object {
