@@ -1,15 +1,18 @@
 package it.cammino.risuscito.dialogs
 
 
-import android.annotation.SuppressLint
 import android.app.Dialog
-import android.content.DialogInterface
 import android.os.Bundle
 import android.view.KeyEvent
 import androidx.annotation.ArrayRes
 import androidx.annotation.StringRes
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.os.bundleOf
 import androidx.fragment.app.DialogFragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.viewModels
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import com.afollestad.materialdialogs.MaterialDialog
 import com.afollestad.materialdialogs.list.listItemsSingleChoice
 import java.io.Serializable
@@ -17,33 +20,20 @@ import java.io.Serializable
 @Suppress("unused")
 class ListChoiceDialogFragment : DialogFragment() {
 
-    private var mCallback: ListChoiceCallback? = null
+    private val viewModel: DialogViewModel by viewModels({ requireActivity() })
 
     private val builder: Builder?
         get() = if (arguments?.containsKey(BUILDER_TAG) != true) null else arguments?.getSerializable(BUILDER_TAG) as? Builder
 
-    override fun onDestroyView() {
-        if (retainInstance)
-            dialog?.setDismissMessage(null)
-        super.onDestroyView()
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        retainInstance = true
-    }
-
-    @SuppressLint("CheckResult")
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
         val mBuilder = builder
                 ?: throw IllegalStateException("ListChoiceDialogFragment should be created using its Builder interface.")
 
-        if (mCallback == null)
-            mCallback = mBuilder.mListener
-
         val dialog = MaterialDialog(requireContext())
                 .listItemsSingleChoice(mBuilder.listArrayId, initialSelection = mBuilder.initialSelection) { _, index, _ ->
-                    mCallback?.onPositive(mBuilder.mTag, index)
+                    viewModel.index = index
+                    viewModel.handled = false
+                    viewModel.state.value = DialogState.Positive(this)
                 }
 
         if (mBuilder.mTitle != 0)
@@ -62,7 +52,8 @@ class ListChoiceDialogFragment : DialogFragment() {
 
         mBuilder.mNegativeButton?.let {
             dialog.negativeButton(text = it) {
-                mCallback?.onNegative(mBuilder.mTag)
+                viewModel.handled = false
+                viewModel.state.value = DialogState.Negative(this)
             }
         }
 
@@ -80,35 +71,19 @@ class ListChoiceDialogFragment : DialogFragment() {
         return dialog
     }
 
-    @SuppressLint("CheckResult")
     fun setContent(@StringRes res: Int) {
         (dialog as? MaterialDialog)?.message(res)
-    }
-
-    fun setmCallback(callback: ListChoiceCallback) {
-        mCallback = callback
     }
 
     fun cancel() {
         dialog?.cancel()
     }
 
-    fun setOnCancelListener(listener: DialogInterface.OnCancelListener) {
-        dialog?.setOnCancelListener(listener)
-    }
-
     override fun dismiss() {
         super.dismissAllowingStateLoss()
     }
 
-    override fun onCancel(dialog: DialogInterface) {
-        super.onCancel(dialog)
-        val mBuilder = builder
-        if (mBuilder?.mCanceListener == true)
-            mCallback?.onNegative(mBuilder.mTag)
-    }
-
-    class Builder(context: AppCompatActivity, @field:Transient internal var mListener: ListChoiceCallback, internal val mTag: String) : Serializable {
+    class Builder(context: AppCompatActivity, internal val mTag: String) : Serializable {
 
         @Transient
         private val mContext: AppCompatActivity = context
@@ -118,7 +93,7 @@ class ListChoiceDialogFragment : DialogFragment() {
         internal var mNegativeButton: CharSequence? = null
         internal var mCanceable = false
         internal var mAutoDismiss = true
-        internal var mCanceListener = false
+
         internal var listArrayId: Int = 0
         internal var initialSelection: Int = -1
 
@@ -157,11 +132,6 @@ class ListChoiceDialogFragment : DialogFragment() {
             return this
         }
 
-        fun setHasCancelListener(): Builder {
-            mCanceListener = true
-            return this
-        }
-
         fun setCanceable(): Builder {
             mCanceable = true
             return this
@@ -172,54 +142,34 @@ class ListChoiceDialogFragment : DialogFragment() {
             return this
         }
 
-        fun build(): ListChoiceDialogFragment {
-            val dialog = ListChoiceDialogFragment()
-            val args = Bundle()
-            args.putSerializable(BUILDER_TAG, this)
-            dialog.arguments = args
-            return dialog
-        }
-
-        fun show(): ListChoiceDialogFragment {
-            val dialog = build()
-            if (!mContext.isFinishing)
-                dialog.show(mContext)
-            return dialog
-        }
-    }
-
-    private fun dismissIfNecessary(context: AppCompatActivity, tag: String) {
-        val frag = context.supportFragmentManager.findFragmentByTag(tag)
-        frag?.let {
-            (it as? DialogFragment)?.dismiss()
-            context.supportFragmentManager.beginTransaction()
-                    .remove(it).commit()
-        }
-    }
-
-    fun show(context: AppCompatActivity): ListChoiceDialogFragment {
-        val builder = builder
-        builder?.let {
-            dismissIfNecessary(context, it.mTag)
-            show(context.supportFragmentManager, it.mTag)
-        }
-        return this
-    }
-
-    interface ListChoiceCallback {
-        fun onPositive(tag: String, index: Int)
-        fun onNegative(tag: String)
     }
 
     companion object {
-        private const val BUILDER_TAG = "builder"
-        fun findVisible(context: AppCompatActivity?, tag: String): ListChoiceDialogFragment? {
-            context?.let {
-                val frag = it.supportFragmentManager.findFragmentByTag(tag)
-                return if (frag != null && frag is ListChoiceDialogFragment) frag else null
+
+        private const val BUILDER_TAG = "bundle_builder"
+
+        private fun newInstance() = ListChoiceDialogFragment()
+
+        private fun newInstance(builder: Builder): ListChoiceDialogFragment {
+            return newInstance().apply {
+                arguments = bundleOf(
+                        Pair(BUILDER_TAG, builder)
+                )
             }
-            return null
         }
+
+        fun show(builder: Builder, fragmentManger: FragmentManager) {
+            newInstance(builder).run {
+                show(fragmentManger, builder.mTag)
+            }
+        }
+
+    }
+
+    class DialogViewModel : ViewModel() {
+        var index: Int = 0
+        var handled = true
+        val state = MutableLiveData<DialogState<ListChoiceDialogFragment>>()
     }
 
 }
