@@ -19,7 +19,6 @@ package it.cammino.risuscito.playback
 import android.annotation.SuppressLint
 import android.content.Context
 import android.graphics.BitmapFactory
-import android.os.AsyncTask
 import android.support.v4.media.MediaMetadataCompat
 import android.util.Log
 import it.cammino.risuscito.LUtils
@@ -31,6 +30,9 @@ import it.cammino.risuscito.Utility.retrieveMediaFileLink
 import it.cammino.risuscito.database.RisuscitoDatabase
 import it.cammino.risuscito.database.dao.CantoDao
 import it.cammino.risuscito.ui.RisuscitoApplication
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
 import java.util.*
 
 @Suppress("unused")
@@ -45,6 +47,7 @@ class MusicProvider internal constructor(private val mContext: Context) {
 
     // Categorized caches for music track data:
     private val mMusicListById: LinkedHashMap<String, MediaMetadataCompat> = LinkedHashMap()
+
     @Volatile
     private var mCurrentState = State.NON_INITIALIZED
     private val mDao: CantoDao = RisuscitoDatabase.getInstance(mContext).cantoDao()
@@ -95,18 +98,10 @@ class MusicProvider internal constructor(private val mContext: Context) {
             callback?.onMusicCatalogReady(true)
             return
         }
-
-        // Asynchronously load the music catalog in a separate thread
-        object : AsyncTask<Void, Void, State>() {
-            override fun doInBackground(vararg params: Void): State {
-                retrieveMedia()
-                return mCurrentState
-            }
-
-            override fun onPostExecute(current: State) {
-                callback?.onMusicCatalogReady(current == State.INITIALIZED)
-            }
-        }.execute()
+        GlobalScope.launch(Dispatchers.IO) {
+            retrieveMedia()
+            callback?.onMusicCatalogReady(mCurrentState == State.INITIALIZED)
+        }
     }
 
     @Synchronized
@@ -118,28 +113,20 @@ class MusicProvider internal constructor(private val mContext: Context) {
 //            Log.d(TAG, "LINGUA PREFERENCE: " + PreferenceManager.getDefaultSharedPreferences(mContext).getString(Utility.SYSTEM_LANGUAGE, ""))
 
             var mNewBase = mContext
-            mNewBase = RisuscitoApplication.localeManager.setLocale(mNewBase)
+            mNewBase = RisuscitoApplication.localeManager.useCustomConfig(mNewBase)
 
             val art = decodeSampledBitmapFromResource(mNewBase.resources, R.drawable.ic_launcher_144dp, 320, 320)
             val artSmall = BitmapFactory.decodeResource(mNewBase.resources, R.mipmap.ic_launcher)
 
             val canti = mDao.allByWithLink
-            Log.d(TAG, "retrieveMedia: " + canti.size)
+            Log.d(TAG, "$RETRIEVE_MEDIA: ${canti.size}")
 
             var temp: MediaMetadataCompat
 
             for (canto in canti) {
-                Log.d(
-                        TAG,
-                        "retrieveMedia: "
-                                + canto.id
-                                + " / "
-                                + mNewBase.resources.getString(LUtils.getResId(canto.titolo, R.string::class.java))
-                                + " / "
-                                + if (LUtils.getResId(canto.link, R.string::class.java) != -1) mNewBase.resources.getString(LUtils.getResId(canto.link, R.string::class.java)) else canto.link)
-
+                Log.d(TAG,
+                        "$RETRIEVE_MEDIA: ${canto.id} / ${mNewBase.resources.getString(LUtils.getResId(canto.titolo, R.string::class.java))} / ${if (LUtils.getResId(canto.link, R.string::class.java) != -1) mNewBase.resources.getString(LUtils.getResId(canto.link, R.string::class.java)) else canto.link}")
                 var url = if (LUtils.getResId(canto.link, R.string::class.java) != -1) mNewBase.resources.getString(LUtils.getResId(canto.link, R.string::class.java)) else canto.link
-//                if (EasyPermissions.hasPermissions(mContext, Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
                 if (isExternalStorageReadable && isDefaultLocationPublic(mNewBase)) {
                     // ho il permesso di scrivere la memoria esterna, quindi cerco il file anche lì
                     if (retrieveMediaFileLink(mContext, url, true).isNotEmpty())
@@ -149,7 +136,7 @@ class MusicProvider internal constructor(private val mContext: Context) {
                         url = retrieveMediaFileLink(mContext, url, false)
                 }
 
-                Log.v(TAG, "retrieveMedia: " + canto.id + " / " + mNewBase.resources.getString(LUtils.getResId(canto.titolo, R.string::class.java)) + " / " + url)
+                Log.v(TAG, "$RETRIEVE_MEDIA: ${canto.id} / ${mNewBase.resources.getString(LUtils.getResId(canto.titolo, R.string::class.java))} / $url")
 
                 if (!url.isNullOrEmpty()) {
                     temp = MediaMetadataCompat.Builder()
@@ -207,6 +194,7 @@ class MusicProvider internal constructor(private val mContext: Context) {
 
         const val MEDIA_ID_ROOT = "__ROOT__"
         const val MEDIA_ID_EMPTY_ROOT = "__EMPTY__"
+        const val RETRIEVE_MEDIA = "retrieveMedia"
         private val TAG = MusicProvider::class.java.simpleName
     }
 }
