@@ -25,6 +25,7 @@ import android.content.IntentFilter
 import android.media.AudioManager
 import android.os.Bundle
 import android.os.Handler
+import android.os.Looper
 import android.os.SystemClock
 import android.support.v4.media.MediaBrowserCompat
 import android.support.v4.media.MediaMetadataCompat
@@ -52,14 +53,14 @@ class MusicService : MediaBrowserServiceCompat() {
     private var mCurrentMedia: MediaSessionCompat.QueueItem? = null
     private var mAudioBecomingNoisyReceiver: AudioBecomingNoisyReceiver? = null
 
-    private val mDelayedStopHandler = Handler(Handler.Callback { msg ->
+    private val mDelayedStopHandler = Handler(Looper.getMainLooper()) { msg ->
         if (msg.what == STOP_CMD && mPlayback?.isPlaying != true) {
             Log.d(TAG, "Stopping service")
             stopSelf()
             mServiceStarted = false
         }
         false
-    })
+    }
 
     override fun onCreate() {
         super.onCreate()
