@@ -1,89 +1,100 @@
 package it.cammino.risuscito.ui
 
-import android.annotation.SuppressLint
 import android.annotation.TargetApi
 import android.content.Context
-import android.content.SharedPreferences
 import android.content.res.Configuration
 import android.content.res.Resources
 import android.os.Build.VERSION_CODES.JELLY_BEAN_MR1
 import android.os.Build.VERSION_CODES.N
 import android.provider.Settings
 import android.util.Log
+import androidx.annotation.RequiresApi
 import androidx.core.content.edit
 import androidx.preference.PreferenceManager
 import it.cammino.risuscito.LUtils
 import it.cammino.risuscito.Utility
 import java.util.*
 
+
 class LocaleManager(context: Context) {
 
-    private val prefs: SharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
+    //    private val prefs: SharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
+    private var customScale = 0F
 
-    val language: String
-        get() = prefs.getString(Utility.SYSTEM_LANGUAGE, "") ?: ""
+    init {
+        Log.d(TAG, "init language: ${getLanguage(context)}")
 
-    fun setLocale(c: Context): Context {
-        return updateResources(c, language)
-    }
-
-    @SuppressLint("ApplySharedPref")
-    fun persistLanguage(language: String) {
-        prefs.edit {
-            putString(Utility.SYSTEM_LANGUAGE, language)
-        }
-    }
-
-    @Suppress("DEPRECATION")
-    private fun updateResources(context: Context, language: String): Context {
-        val res = context.resources
-        val config = Configuration(res.configuration)
-        lateinit var locale: Locale
-
-        Log.d(TAG, "language: $language")
-
-        if (language.isNotEmpty()) {
-            Log.d(TAG, "attachBaseContext - settings language set: $language")
-            locale = Locale(language)
-            Locale.setDefault(locale)
+        if (getLanguage(context).isNotEmpty()) {
+            Log.d(TAG, "init - settings language set: ${getLanguage(context)}")
         } else {
-            val mLanguage = when (getSystemLocale(res).language) {
+            // non è ancora stata impostata nessuna lingua nelle impostazioni --> setto una lingua
+            // selezionabile oppure IT se non presente
+            val mLanguage = when (getSystemLocale(context.resources).language) {
                 LANGUAGE_UKRAINIAN -> LANGUAGE_UKRAINIAN
-                LANGUAGE_ENGLISH -> LANGUAGE_ENGLISH
+                LANGUAGE_ENGLISH -> if (getSystemLocale(context.resources).country.isNotEmpty() && getSystemLocale(context.resources).country == COUNTRY_PHILIPPINES)
+                    LANGUAGE_ENGLISH_PHILIPPINES
+                else
+                    LANGUAGE_ENGLISH
                 LANGUAGE_TURKISH -> LANGUAGE_TURKISH
+                LANGUAGE_POLISH -> LANGUAGE_POLISH
                 else -> LANGUAGE_ITALIAN
             }
             Log.d(TAG, "attachBaseContext - default language set: $mLanguage")
-            persistLanguage(mLanguage)
-            locale = Locale(mLanguage)
-            Locale.setDefault(locale)
-        }// non è ancora stata impostata nessuna lingua nelle impostazioni --> setto una lingua
-        // selezionabile oppure IT se non presente
+            persistLanguage(context, mLanguage)
+        }
 
-        // font dimension
+        var returnScale = 0F
         try {
-            val actualScale = config.fontScale
+            val actualScale = context.resources.configuration.fontScale
             Log.d(TAG, "actualScale: $actualScale")
             val systemScale = Settings.System.getFloat(context.contentResolver, Settings.System.FONT_SCALE)
             Log.d(TAG, "systemScale: $systemScale")
             if (actualScale != systemScale) {
-                config.fontScale = systemScale
+                returnScale = systemScale
             }
         } catch (e: Settings.SettingNotFoundException) {
             Log.w(TAG, "Settings.SettingNotFoundException - FUNZIONE RESIZE TESTO NON SUPPORTATA: ${e.localizedMessage}")
         } catch (e: NullPointerException) {
             Log.w(TAG, "Settings.SettingNotFoundException - FUNZIONE RESIZE TESTO NON SUPPORTATA: ${e.localizedMessage}")
         }
+        customScale = returnScale
 
-        setSystemLocale(config, locale)
-        return if (LUtils.hasJB()) {
-            config.setLayoutDirection(locale)
-            res.updateConfiguration(config, res.displayMetrics)
-            context.createConfigurationContext(config)
-        } else {
-            res.updateConfiguration(config, res.displayMetrics)
-            context
+    }
+
+//    val language: String
+//        get() = prefs.getString(Utility.SYSTEM_LANGUAGE, "") ?: ""
+
+    fun persistLanguage(context: Context, language: String) {
+        PreferenceManager.getDefaultSharedPreferences(context).edit {
+            putString(Utility.SYSTEM_LANGUAGE, language)
         }
+    }
+
+    fun getLanguage(context: Context): String {
+        return PreferenceManager.getDefaultSharedPreferences(context).getString(Utility.SYSTEM_LANGUAGE, "")
+                ?: ""
+    }
+
+    fun useCustomConfig(context: Context): Context {
+        Log.d(TAG, "useCustomConfig language: ${getLanguage(context)}")
+        val locale = if (getLanguage(context) == LANGUAGE_ENGLISH_PHILIPPINES) Locale(LANGUAGE_ENGLISH, COUNTRY_PHILIPPINES) else Locale(getLanguage(context))
+        Log.d(TAG, "useCustomConfig language: ${locale.language}")
+        Log.d(TAG, "useCustomConfig country: ${locale.country}")
+        Locale.setDefault(locale)
+        val config = if (LUtils.hasJB()) Configuration() else Configuration(context.resources.configuration)
+        if (customScale > 0F)
+            config.fontScale = customScale
+        setSystemLocale(config, locale)
+        return context.updateConfiguration(config)
+    }
+
+    fun updateConfigurationIfSupported(context: Context, overrideConfiguration: Configuration?): Configuration? {
+        overrideConfiguration?.let { config ->
+            if (isLocaleNotEmpty(config))
+                return config
+            setSystemLocale(config, if (getLanguage(context) == LANGUAGE_ENGLISH_PHILIPPINES) Locale(LANGUAGE_ENGLISH, COUNTRY_PHILIPPINES) else Locale(getLanguage(context)))
+        }
+        return overrideConfiguration
     }
 
     companion object {
@@ -93,6 +104,45 @@ class LocaleManager(context: Context) {
         const val LANGUAGE_ENGLISH = "en"
         const val LANGUAGE_UKRAINIAN = "uk"
         const val LANGUAGE_TURKISH = "tr"
+        const val LANGUAGE_POLISH = "pl"
+        const val LANGUAGE_ENGLISH_PHILIPPINES = "ep"
+
+        const val COUNTRY_PHILIPPINES = "PH"
+
+        @Suppress("DEPRECATION")
+        private fun updateConfigurationLegacy(context: Context, config: Configuration): Context {
+            context.resources.updateConfiguration(config, context.resources.displayMetrics)
+            return context
+        }
+
+        @RequiresApi(JELLY_BEAN_MR1)
+        private fun updateConfigurationJB(context: Context, config: Configuration): Context {
+            return context.createConfigurationContext(config)
+        }
+
+        fun Context.updateConfiguration(config: Configuration): Context {
+            return if (LUtils.hasJB())
+                updateConfigurationJB(this, config)
+            else
+                updateConfigurationLegacy(this, config)
+        }
+
+        @Suppress("DEPRECATION")
+        private fun isLocaleNotEmptyLegacy(config: Configuration): Boolean {
+            return config.locale != null
+        }
+
+        @TargetApi(N)
+        private fun isLocaleNotEmptyN(config: Configuration): Boolean {
+            return !config.locales.isEmpty
+        }
+
+        fun isLocaleNotEmpty(config: Configuration): Boolean {
+            return if (LUtils.hasN())
+                isLocaleNotEmptyN(config)
+            else
+                isLocaleNotEmptyLegacy(config)
+        }
 
         @Suppress("DEPRECATION")
         private fun getSystemLocaleLegacy(config: Configuration): Locale {
