@@ -1,7 +1,6 @@
 package it.cammino.risuscito.ui
 
 import android.content.Context
-import android.content.res.Configuration
 import android.graphics.drawable.Drawable
 import android.net.Uri
 import android.widget.ImageView
@@ -11,7 +10,9 @@ import com.mikepenz.materialdrawer.util.DrawerImageLoader
 import com.squareup.picasso.Picasso
 import it.cammino.risuscito.database.RisuscitoDatabase
 import it.cammino.risuscito.utils.ThemeUtils
-import it.cammino.risuscito.utils.ioThread
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
 
 
 class RisuscitoApplication : MultiDexApplication() {
@@ -21,12 +22,8 @@ class RisuscitoApplication : MultiDexApplication() {
 
         ThemeUtils.setDefaultNightMode(applicationContext)
 
-        //only required if you add a custom or generic font on your own
-//        Iconics.init(applicationContext)
-
-        ioThread {
-            RisuscitoDatabase.getInstance(this).cantoDao().getCantoById(1)
-        }
+        val mDao = RisuscitoDatabase.getInstance(this).cantoDao()
+        GlobalScope.launch(Dispatchers.IO) { mDao.getCantoById(1) }
 
         // initialize and create the image loader logic
         DrawerImageLoader.init(
@@ -41,18 +38,13 @@ class RisuscitoApplication : MultiDexApplication() {
                 })
     }
 
-    override fun onConfigurationChanged(newConfig: Configuration) {
-        super.onConfigurationChanged(newConfig)
-        localeManager.setLocale(this)
-    }
-
     override fun attachBaseContext(base: Context) {
         localeManager = LocaleManager(base)
-        super.attachBaseContext(localeManager.setLocale(base))
+        super.attachBaseContext(localeManager.useCustomConfig(base))
     }
 
     companion object {
-        //        private val TAG = RisuscitoApplication::class.java.canonicalName
+        internal val TAG = RisuscitoApplication::class.java.canonicalName
         lateinit var localeManager: LocaleManager
     }
 }
